@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -12,29 +12,27 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import os
-import re
 
 from codecs import open
-from datetime import datetime, timezone
 from os.path import dirname, join, realpath
 
 from setuptools import find_packages, setup
 
-DESCRIPTION = "Probabilistic Programming in Python: Bayesian Modeling and Probabilistic Machine Learning with Aesara"
+import versioneer
+
+DESCRIPTION = "Probabilistic Programming in Python: Bayesian Modeling and Probabilistic Machine Learning with PyTensor"
 AUTHOR = "PyMC Developers"
 AUTHOR_EMAIL = "pymc.devs@gmail.com"
 URL = "http://github.com/pymc-devs/pymc"
 LICENSE = "Apache License, Version 2.0"
-NIGHLTY = "BUILD_PYMC_NIGHTLY" in os.environ
 
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
     "License :: OSI Approved :: Apache Software License",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
@@ -55,37 +53,11 @@ with open(REQUIREMENTS_FILE) as f:
 
 test_reqs = ["pytest", "pytest-cov"]
 
-
-def get_distname(nightly_build=False):
-    distname = "pymc"
-    if nightly_build:
-        distname = f"{distname}-nightly"
-
-    return distname
-
-
-def get_version(nightly_build=False):
-    version_file = join("pymc", "__init__.py")
-    lines = open(version_file).readlines()
-    version_regex = r"^__version__ = ['\"]([^'\"]*)['\"]"
-    for line in lines:
-        mo = re.search(version_regex, line, re.M)
-        if mo:
-            version = mo.group(1)
-
-            if nightly_build:
-                suffix = datetime.now(timezone.utc).strftime(r".dev%Y%m%d")
-                version = f"{version}{suffix}"
-
-            return version
-
-    raise RuntimeError(f"Unable to find version in {version_file}.")
-
-
 if __name__ == "__main__":
     setup(
-        name=get_distname(NIGHLTY),
-        version=get_version(NIGHLTY),
+        name="pymc",
+        version=versioneer.get_version(),
+        cmdclass=versioneer.get_cmdclass(),
         maintainer=AUTHOR,
         maintainer_email=AUTHOR_EMAIL,
         description=DESCRIPTION,
@@ -93,13 +65,12 @@ if __name__ == "__main__":
         url=URL,
         long_description=LONG_DESCRIPTION,
         long_description_content_type="text/x-rst",
-        packages=find_packages(),
+        packages=find_packages(exclude=["tests*"]),
         # because of an upload-size limit by PyPI, we're temporarily removing docs from the tarball.
         # Also see MANIFEST.in
         # package_data={'docs': ['*']},
-        include_package_data=True,
         classifiers=classifiers,
-        python_requires=">=3.7",
+        python_requires=">=3.9",
         install_requires=install_reqs,
         tests_require=test_reqs,
     )
