@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -60,7 +60,87 @@ Loading a saved backend
 Saved backends can be loaded using `arviz.from_netcdf`
 
 """
+from copy import copy
+from typing import Dict, List, Mapping, Optional, Sequence, Tuple, Union
+
+import numpy as np
+
+from typing_extensions import TypeAlias
+
 from pymc.backends.arviz import predictions_to_inference_data, to_inference_data
-from pymc.backends.ndarray import NDArray, point_list_to_multitrace
+from pymc.backends.base import BaseTrace, IBaseTrace
+from pymc.backends.ndarray import NDArray
+from pymc.model import Model
+from pymc.step_methods.compound import BlockedStep, CompoundStep
+
+HAS_MCB = False
+try:
+    from mcbackend import Backend, Run
+
+    from pymc.backends.mcbackend import init_chain_adapters
+
+    TraceOrBackend = Union[BaseTrace, Backend]
+    RunType: TypeAlias = Run
+    HAS_MCB = True
+except ImportError:
+    TraceOrBackend = BaseTrace  # type: ignore
+    RunType = type(None)  # type: ignore
+
 
 __all__ = ["to_inference_data", "predictions_to_inference_data"]
+
+
+def _init_trace(
+    *,
+    expected_length: int,
+    chain_number: int,
+    stats_dtypes: List[Dict[str, type]],
+    trace: Optional[BaseTrace],
+    model: Model,
+) -> BaseTrace:
+    """Initializes a trace backend for a chain."""
+    strace: BaseTrace
+    if trace is None:
+        strace = NDArray(model=model)
+    elif isinstance(trace, BaseTrace):
+        if len(trace) > 0:
+            raise ValueError("Continuation of traces is no longer supported.")
+        strace = copy(trace)
+    else:
+        raise NotImplementedError(f"Unsupported `trace`: {trace}")
+
+    strace.setup(expected_length, chain_number, stats_dtypes)
+    return strace
+
+
+def init_traces(
+    *,
+    backend: Optional[TraceOrBackend],
+    chains: int,
+    expected_length: int,
+    step: Union[BlockedStep, CompoundStep],
+    initial_point: Mapping[str, np.ndarray],
+    model: Model,
+) -> Tuple[Optional[RunType], Sequence[IBaseTrace]]:
+    """Initializes a trace recorder for each chain."""
+    if HAS_MCB and isinstance(backend, Backend):
+        return init_chain_adapters(
+            backend=backend,
+            chains=chains,
+            initial_point=initial_point,
+            step=step,
+            model=model,
+        )
+
+    assert backend is None or isinstance(backend, BaseTrace)
+    traces = [
+        _init_trace(
+            expected_length=expected_length,
+            stats_dtypes=step.stats_dtypes,
+            chain_number=chain_number,
+            trace=backend,
+            model=model,
+        )
+        for chain_number in range(chains)
+    ]
+    return None, traces
