@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -12,34 +12,36 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import collections
 import io
-import os
-import pkgutil
 import urllib.request
 import warnings
 
 from copy import copy
-from typing import Any, Dict, List, Optional, Sequence, Union, cast
+from typing import Dict, Optional, Sequence, Tuple, Union, cast
 
-import aesara
-import aesara.tensor as at
 import numpy as np
+import pandas as pd
+import pytensor
+import pytensor.tensor as pt
+import xarray as xr
 
-from aesara.compile.sharedvalue import SharedVariable
-from aesara.graph.basic import Apply
-from aesara.tensor.type import TensorType
-from aesara.tensor.var import TensorConstant, TensorVariable
+from pytensor.compile.sharedvalue import SharedVariable
+from pytensor.raise_op import Assert
+from pytensor.scalar import Cast
+from pytensor.tensor.elemwise import Elemwise
+from pytensor.tensor.random.basic import IntegersRV
+from pytensor.tensor.subtensor import AdvancedSubtensor
+from pytensor.tensor.type import TensorType
+from pytensor.tensor.var import TensorConstant, TensorVariable
 
 import pymc as pm
 
-from pymc.aesaraf import pandas_to_array
+from pymc.pytensorf import convert_observed_data
 
 __all__ = [
     "get_data",
     "GeneratorAdapter",
     "Minibatch",
-    "align_minibatches",
     "Data",
     "ConstantData",
     "MutableData",
@@ -59,18 +61,14 @@ def get_data(filename):
     -------
     BytesIO of the data
     """
-    data_pkg = "pymc.tests"
-    try:
-        content = pkgutil.get_data(data_pkg, os.path.join("data", filename))
-    except FileNotFoundError:
-        with urllib.request.urlopen(BASE_URL.format(filename=filename)) as handle:
-            content = handle.read()
+    with urllib.request.urlopen(BASE_URL.format(filename=filename)) as handle:
+        content = handle.read()
     return io.BytesIO(content)
 
 
 class GenTensorVariable(TensorVariable):
     def __init__(self, op, type, name=None):
-        super().__init__(type=type, name=name)
+        super().__init__(type=type, owner=None, name=name)
         self.op = op
 
     def set_gen(self, gen):
@@ -126,353 +124,95 @@ class GeneratorAdapter:
         return hash(id(self))
 
 
-class Minibatch(TensorVariable):
-    """Multidimensional minibatch that is pure TensorVariable
+class MinibatchIndexRV(IntegersRV):
+    _print_name = ("minibatch_index", r"\operatorname{minibatch\_index}")
+
+    # Work-around for https://github.com/pymc-devs/pytensor/issues/97
+    def make_node(self, rng, *args, **kwargs):
+        if rng is None:
+            rng = pytensor.shared(np.random.default_rng())
+        return super().make_node(rng, *args, **kwargs)
+
+
+minibatch_index = MinibatchIndexRV()
+
+
+def is_minibatch(v: TensorVariable) -> bool:
+    return (
+        isinstance(v.owner.op, AdvancedSubtensor)
+        and isinstance(v.owner.inputs[1].owner.op, MinibatchIndexRV)
+        and valid_for_minibatch(v.owner.inputs[0])
+    )
+
+
+def valid_for_minibatch(v: TensorVariable) -> bool:
+    return (
+        v.owner is None
+        # The only PyTensor operation we allow on observed data is type casting
+        # Although we could allow for any graph that does not depend on other RVs
+        or (
+            isinstance(v.owner.op, Elemwise)
+            and v.owner.inputs[0].owner is None
+            and isinstance(v.owner.op.scalar_op, Cast)
+        )
+    )
+
+
+def assert_all_scalars_equal(scalar, *scalars):
+    if len(scalars) == 0:
+        return scalar
+    else:
+        return Assert(
+            "All variables shape[0] in Minibatch should be equal, check your Minibatch(data1, data2, ...) code"
+        )(scalar, pt.all([pt.eq(scalar, s) for s in scalars]))
+
+
+def Minibatch(variable: TensorVariable, *variables: TensorVariable, batch_size: int):
+    """Get random slices from variables from the leading dimension.
 
     Parameters
     ----------
-    data: np.ndarray
-        initial data
-    batch_size: ``int`` or ``List[int|tuple(size, random_seed)]``
-        batch size for inference, random seed is needed
-        for child random generators
-    dtype: ``str``
-        cast data to specific type
-    broadcastable: tuple[bool]
-        change broadcastable pattern that defaults to ``(False, ) * ndim``
-    name: ``str``
-        name for tensor, defaults to "Minibatch"
-    random_seed: ``int``
-        random seed that is used by default
-    update_shared_f: ``callable``
-        returns :class:`ndarray` that will be carefully
-        stored to underlying shared variable
-        you can use it to change source of
-        minibatches programmatically
-    in_memory_size: ``int`` or ``List[int|slice|Ellipsis]``
-        data size for storing in ``aesara.shared``
-
-    Attributes
-    ----------
-    shared: shared tensor
-        Used for storing data
-    minibatch: minibatch tensor
-        Used for training
-
-    Notes
-    -----
-    Below is a common use case of Minibatch with variational inference.
-    Importantly, we need to make PyMC "aware" that a minibatch is being used in inference.
-    Otherwise, we will get the wrong :math:`logp` for the model.
-    the density of the model ``logp`` that is affected by Minibatch. See more in the examples below.
-    To do so, we need to pass the ``total_size`` parameter to the observed node, which correctly scales
-    the density of the model ``logp`` that is affected by Minibatch. See more in the examples below.
+    variable: TensorVariable
+    variables: TensorVariable
+    batch_size: int
 
     Examples
     --------
-    Consider we have `data` as follows:
-
-    >>> data = np.random.rand(100, 100)
-
-    if we want a 1d slice of size 10 we do
-
-    >>> x = Minibatch(data, batch_size=10)
-
-    Note that your data is cast to ``floatX`` if it is not integer type
-    But you still can add the ``dtype`` kwarg for :class:`Minibatch`
-    if you need more control.
-
-    If we want 10 sampled rows and columns
-    ``[(size, seed), (size, seed)]`` we can use
-
-    >>> x = Minibatch(data, batch_size=[(10, 42), (10, 42)], dtype='int32')
-    >>> assert str(x.dtype) == 'int32'
-
-
-    Or, more simply, we can use the default random seed = 42
-    ``[size, size]``
-
-    >>> x = Minibatch(data, batch_size=[10, 10])
-
-
-    In the above, `x` is a regular :class:`TensorVariable` that supports any math operations:
-
-
-    >>> assert x.eval().shape == (10, 10)
-
-
-    You can pass the Minibatch `x` to your desired model:
-
-    >>> with pm.Model() as model:
-    ...     mu = pm.Flat('mu')
-    ...     sd = pm.HalfNormal('sd')
-    ...     lik = pm.Normal('lik', mu, sd, observed=x, total_size=(100, 100))
-
-
-    Then you can perform regular Variational Inference out of the box
-
-
-    >>> with model:
-    ...     approx = pm.fit()
-
-
-    Important note: :class:``Minibatch`` has ``shared``, and ``minibatch`` attributes
-    you can call later:
-
-    >>> x.set_value(np.random.laplace(size=(100, 100)))
-
-    and minibatches will be then from new storage
-    it directly affects ``x.shared``.
-    A less convenient convenient, but more explicit, way to achieve the same
-    thing:
-
-    >>> x.shared.set_value(pm.floatX(np.random.laplace(size=(100, 100))))
-
-    The programmatic way to change storage is as follows
-    I import ``partial`` for simplicity
-    >>> from functools import partial
-    >>> datagen = partial(np.random.laplace, size=(100, 100))
-    >>> x = Minibatch(datagen(), batch_size=10, update_shared_f=datagen)
-    >>> x.update_shared()
-
-    To be more concrete about how we create a minibatch, here is a demo:
-    1. create a shared variable
-
-        >>> shared = aesara.shared(data)
-
-    2. take a random slice of size 10:
-
-        >>> ridx = pm.at_rng().uniform(size=(10,), low=0, high=data.shape[0]-1e-10).astype('int64')
-
-    3) take the resulting slice:
-
-        >>> minibatch = shared[ridx]
-
-    That's done. Now you can use this minibatch somewhere else.
-    You can see that the implementation does not require a fixed shape
-    for the shared variable. Feel free to use that if needed.
-    *FIXME: What is "that" which we can use here?  A fixed shape?  Should this say
-    "but feel free to put a fixed shape on the shared variable, if appropriate?"*
-
-    Suppose you need to make some replacements in the graph, e.g. change the minibatch to testdata
-
-    >>> node = x ** 2  # arbitrary expressions on minibatch `x`
-    >>> testdata = pm.floatX(np.random.laplace(size=(1000, 10)))
-
-    Then you should create a `dict` with replacements:
-
-    >>> replacements = {x: testdata}
-    >>> rnode = aesara.clone_replace(node, replacements)
-    >>> assert (testdata ** 2 == rnode.eval()).all()
-
-    *FIXME: In the following, what is the **reason** to replace the Minibatch variable with
-    its shared variable?  And in the following, the `rnode` is a **new** node, not a modification
-    of a previously existing node, correct?*
-    To replace a minibatch with its shared variable you should do
-    the same things. The Minibatch variable is accessible through the `minibatch` attribute.
-    For example
-
-    >>> replacements = {x.minibatch: x.shared}
-    >>> rnode = aesara.clone_replace(node, replacements)
-
-    For more complex slices some more code is needed that can seem not so clear
-
-    >>> moredata = np.random.rand(10, 20, 30, 40, 50)
-
-    The default ``total_size`` that can be passed to PyMC random node
-    is then ``(10, 20, 30, 40, 50)`` but can be less verbose in some cases
-
-    1. Advanced indexing, ``total_size = (10, Ellipsis, 50)``
-
-        >>> x = Minibatch(moredata, [2, Ellipsis, 10])
-
-        We take the slice only for the first and last dimension
-
-        >>> assert x.eval().shape == (2, 20, 30, 40, 10)
-
-    2. Skipping a particular dimension, ``total_size = (10, None, 30)``:
-
-        >>> x = Minibatch(moredata, [2, None, 20])
-        >>> assert x.eval().shape == (2, 20, 20, 40, 50)
-
-    3. Mixing both of these together, ``total_size = (10, None, 30, Ellipsis, 50)``:
-
-        >>> x = Minibatch(moredata, [2, None, 20, Ellipsis, 10])
-        >>> assert x.eval().shape == (2, 20, 20, 40, 10)
+    >>> data1 = np.random.randn(100, 10)
+    >>> data2 = np.random.randn(100, 20)
+    >>> mdata1, mdata2 = Minibatch(data1, data2, batch_size=10)
     """
 
-    RNG = collections.defaultdict(list)  # type: Dict[str, List[Any]]
+    if not isinstance(batch_size, int):
+        raise TypeError("batch_size must be an integer")
 
-    @aesara.config.change_flags(compute_test_value="raise")
-    def __init__(
-        self,
-        data,
-        batch_size=128,
-        dtype=None,
-        broadcastable=None,
-        name="Minibatch",
-        random_seed=42,
-        update_shared_f=None,
-        in_memory_size=None,
-    ):
-        if dtype is None:
-            data = pm.smartfloatX(np.asarray(data))
-        else:
-            data = np.asarray(data, dtype)
-        in_memory_slc = self.make_static_slices(in_memory_size)
-        self.shared = aesara.shared(data[tuple(in_memory_slc)])
-        self.update_shared_f = update_shared_f
-        self.random_slc = self.make_random_slices(self.shared.shape, batch_size, random_seed)
-        minibatch = self.shared[self.random_slc]
-        if broadcastable is None:
-            broadcastable = (False,) * minibatch.ndim
-        minibatch = at.patternbroadcast(minibatch, broadcastable)
-        self.minibatch = minibatch
-        super().__init__(self.minibatch.type, None, None, name=name)
-        Apply(aesara.compile.view_op, inputs=[self.minibatch], outputs=[self])
-        self.tag.test_value = copy(self.minibatch.tag.test_value)
-
-    def rslice(self, total, size, seed):
-        if size is None:
-            return slice(None)
-        elif isinstance(size, int):
-            rng = pm.at_rng(seed)
-            Minibatch.RNG[id(self)].append(rng)
-            return rng.uniform(size=(size,), low=0.0, high=pm.floatX(total) - 1e-16).astype("int64")
-        else:
-            raise TypeError("Unrecognized size type, %r" % size)
-
-    def __del__(self):
-        del Minibatch.RNG[id(self)]
-
-    @staticmethod
-    def make_static_slices(user_size):
-        if user_size is None:
-            return [Ellipsis]
-        elif isinstance(user_size, int):
-            return slice(None, user_size)
-        elif isinstance(user_size, (list, tuple)):
-            slc = list()
-            for i in user_size:
-                if isinstance(i, int):
-                    slc.append(i)
-                elif i is None:
-                    slc.append(slice(None))
-                elif i is Ellipsis:
-                    slc.append(Ellipsis)
-                elif isinstance(i, slice):
-                    slc.append(i)
-                else:
-                    raise TypeError("Unrecognized size type, %r" % user_size)
-            return slc
-        else:
-            raise TypeError("Unrecognized size type, %r" % user_size)
-
-    def make_random_slices(self, in_memory_shape, batch_size, default_random_seed):
-        if batch_size is None:
-            return [Ellipsis]
-        elif isinstance(batch_size, int):
-            slc = [self.rslice(in_memory_shape[0], batch_size, default_random_seed)]
-        elif isinstance(batch_size, (list, tuple)):
-
-            def check(t):
-                if t is Ellipsis or t is None:
-                    return True
-                else:
-                    if isinstance(t, (tuple, list)):
-                        if not len(t) == 2:
-                            return False
-                        else:
-                            return isinstance(t[0], int) and isinstance(t[1], int)
-                    elif isinstance(t, int):
-                        return True
-                    else:
-                        return False
-
-            # end check definition
-            if not all(check(t) for t in batch_size):
-                raise TypeError(
-                    "Unrecognized `batch_size` type, expected "
-                    "int or List[int|tuple(size, random_seed)] where "
-                    "size and random seed are both ints, got %r" % batch_size
-                )
-            batch_size = [(i, default_random_seed) if isinstance(i, int) else i for i in batch_size]
-            shape = in_memory_shape
-            if Ellipsis in batch_size:
-                sep = batch_size.index(Ellipsis)
-                begin = batch_size[:sep]
-                end = batch_size[sep + 1 :]
-                if Ellipsis in end:
-                    raise ValueError(
-                        "Double Ellipsis in `batch_size` is restricted, got %r" % batch_size
-                    )
-                if len(end) > 0:
-                    shp_mid = shape[sep : -len(end)]
-                    mid = [at.arange(s) for s in shp_mid]
-                else:
-                    mid = []
-            else:
-                begin = batch_size
-                end = []
-                mid = []
-            if (len(begin) + len(end)) > len(in_memory_shape.eval()):
-                raise ValueError(
-                    "Length of `batch_size` is too big, "
-                    "number of ints is bigger that ndim, got %r" % batch_size
-                )
-            if len(end) > 0:
-                shp_end = shape[-len(end) :]
-            else:
-                shp_end = np.asarray([])
-            shp_begin = shape[: len(begin)]
-            slc_begin = [
-                self.rslice(shp_begin[i], t[0], t[1]) if t is not None else at.arange(shp_begin[i])
-                for i, t in enumerate(begin)
-            ]
-            slc_end = [
-                self.rslice(shp_end[i], t[0], t[1]) if t is not None else at.arange(shp_end[i])
-                for i, t in enumerate(end)
-            ]
-            slc = slc_begin + mid + slc_end
-        else:
-            raise TypeError("Unrecognized size type, %r" % batch_size)
-        return pm.aesaraf.ix_(*slc)
-
-    def update_shared(self):
-        if self.update_shared_f is None:
-            raise NotImplementedError("No `update_shared_f` was provided to `__init__`")
-        self.set_value(np.asarray(self.update_shared_f(), self.dtype))
-
-    def set_value(self, value):
-        self.shared.set_value(np.asarray(value, self.dtype))
-
-    def clone(self):
-        ret = self.type()
-        ret.name = self.name
-        ret.tag = copy(self.tag)
-        return ret
+    tensor, *tensors = tuple(map(pt.as_tensor, (variable, *variables)))
+    upper = assert_all_scalars_equal(*[t.shape[0] for t in (tensor, *tensors)])
+    slc = minibatch_index(0, upper, size=batch_size)
+    for i, v in enumerate((tensor, *tensors)):
+        if not valid_for_minibatch(v):
+            raise ValueError(
+                f"{i}: {v} is not valid for Minibatch, only constants or constants.astype(dtype) are allowed"
+            )
+    result = tuple([v[slc] for v in (tensor, *tensors)])
+    for i, r in enumerate(result):
+        r.name = f"minibatch.{i}"
+    return result if tensors else result[0]
 
 
-def align_minibatches(batches=None):
-    if batches is None:
-        for rngs in Minibatch.RNG.values():
-            for rng in rngs:
-                rng.seed()
-    else:
-        for b in batches:
-            if not isinstance(b, Minibatch):
-                raise TypeError("{b} is not a Minibatch")
-            for rng in Minibatch.RNG[id(b)]:
-                rng.seed()
-
-
-def determine_coords(model, value, dims: Optional[Sequence[str]] = None) -> Dict[str, Sequence]:
+def determine_coords(
+    model,
+    value: Union[pd.DataFrame, pd.Series, xr.DataArray],
+    dims: Optional[Sequence[Optional[str]]] = None,
+    coords: Optional[Dict[str, Union[Sequence, np.ndarray]]] = None,
+) -> Tuple[Dict[str, Union[Sequence, np.ndarray]], Sequence[Optional[str]]]:
     """Determines coordinate values from data or the model (via ``dims``)."""
-    coords = {}
+    if coords is None:
+        coords = {}
 
+    dim_name = None
     # If value is a df or a series, we interpret the index as coords:
     if hasattr(value, "index"):
-        dim_name = None
         if dims is not None:
             dim_name = dims[0]
         if dim_name is None and value.index.name is not None:
@@ -482,13 +222,19 @@ def determine_coords(model, value, dims: Optional[Sequence[str]] = None) -> Dict
 
     # If value is a df, we also interpret the columns as coords:
     if hasattr(value, "columns"):
-        dim_name = None
         if dims is not None:
             dim_name = dims[1]
         if dim_name is None and value.columns.name is not None:
             dim_name = value.columns.name
         if dim_name is not None:
             coords[dim_name] = value.columns
+
+    if isinstance(value, xr.DataArray):
+        if dims is not None:
+            for dim in dims:
+                dim_name = dim
+                # str is applied because dim entries may be None
+                coords[str(dim_name)] = value[dim].to_numpy()
 
     if isinstance(value, np.ndarray) and dims is not None:
         if len(dims) != value.ndim:
@@ -499,10 +245,14 @@ def determine_coords(model, value, dims: Optional[Sequence[str]] = None) -> Dict
             )
         for size, dim in zip(value.shape, dims):
             coord = model.coords.get(dim, None)
-            if coord is None:
+            if coord is None and dim is not None:
                 coords[dim] = range(size)
 
-    return coords
+    if dims is None:
+        # TODO: Also determine dim names from the index
+        dims = [None] * np.ndim(value)
+
+    return coords, dims
 
 
 def ConstantData(
@@ -510,19 +260,29 @@ def ConstantData(
     value,
     *,
     dims: Optional[Sequence[str]] = None,
+    coords: Optional[Dict[str, Union[Sequence, np.ndarray]]] = None,
     export_index_as_coords=False,
+    infer_dims_and_coords=False,
     **kwargs,
 ) -> TensorConstant:
     """Alias for ``pm.Data(..., mutable=False)``.
 
-    Registers the ``value`` as a :class:`~aesara.tensor.TensorConstant` with the model.
+    Registers the ``value`` as a :class:`~pytensor.tensor.TensorConstant` with the model.
     For more information, please reference :class:`pymc.Data`.
     """
+    if export_index_as_coords:
+        infer_dims_and_coords = export_index_as_coords
+        warnings.warn(
+            "Deprecation warning: 'export_index_as_coords; is deprecated and will be removed in future versions. Please use 'infer_dims_and_coords' instead.",
+            DeprecationWarning,
+        )
+
     var = Data(
         name,
         value,
         dims=dims,
-        export_index_as_coords=export_index_as_coords,
+        coords=coords,
+        infer_dims_and_coords=infer_dims_and_coords,
         mutable=False,
         **kwargs,
     )
@@ -534,19 +294,29 @@ def MutableData(
     value,
     *,
     dims: Optional[Sequence[str]] = None,
+    coords: Optional[Dict[str, Union[Sequence, np.ndarray]]] = None,
     export_index_as_coords=False,
+    infer_dims_and_coords=False,
     **kwargs,
 ) -> SharedVariable:
     """Alias for ``pm.Data(..., mutable=True)``.
 
-    Registers the ``value`` as a :class:`~aesara.compile.sharedvalue.SharedVariable`
+    Registers the ``value`` as a :class:`~pytensor.compile.sharedvalue.SharedVariable`
     with the model. For more information, please reference :class:`pymc.Data`.
     """
+    if export_index_as_coords:
+        infer_dims_and_coords = export_index_as_coords
+        warnings.warn(
+            "Deprecation warning: 'export_index_as_coords; is deprecated and will be removed in future versions. Please use 'infer_dims_and_coords' instead.",
+            DeprecationWarning,
+        )
+
     var = Data(
         name,
         value,
         dims=dims,
-        export_index_as_coords=export_index_as_coords,
+        coords=coords,
+        infer_dims_and_coords=infer_dims_and_coords,
         mutable=True,
         **kwargs,
     )
@@ -558,19 +328,26 @@ def Data(
     value,
     *,
     dims: Optional[Sequence[str]] = None,
+    coords: Optional[Dict[str, Union[Sequence, np.ndarray]]] = None,
     export_index_as_coords=False,
+    infer_dims_and_coords=False,
     mutable: Optional[bool] = None,
     **kwargs,
 ) -> Union[SharedVariable, TensorConstant]:
     """Data container that registers a data variable with the model.
 
     Depending on the ``mutable`` setting (default: True), the variable
-    is registered as a :class:`~aesara.compile.sharedvalue.SharedVariable`,
+    is registered as a :class:`~pytensor.compile.sharedvalue.SharedVariable`,
     enabling it to be altered in value and shape, but NOT in dimensionality using
     :func:`pymc.set_data`.
 
     To set the value of the data container variable, check out
-    :func:`pymc.Model.set_data`.
+    :meth:`pymc.Model.set_data`.
+
+    When making predictions or doing posterior predictive sampling, the shape of the
+    registered data variable will most likely need to be changed.  If you encounter an
+    PyTensor shape mismatch error, refer to the documentation for
+    :meth:`pymc.model.set_data`.
 
     For more information, read the notebook :ref:`nb:data_container`.
 
@@ -588,12 +365,16 @@ def Data(
         :ref:`arviz:quickstart`.
         If this parameter is not specified, the random variables will not have dimension
         names.
-    export_index_as_coords : bool, default=False
-        If True, the ``Data`` container will try to infer what the coordinates should be
-        if there is an index in ``value``.
+    coords : dict, optional
+        Coordinate values to set for new dimensions introduced by this ``Data`` variable.
+    export_index_as_coords : bool
+        Deprecated, previous version of "infer_dims_and_coords"
+    infer_dims_and_coords : bool, default=False
+        If True, the ``Data`` container will try to infer what the coordinates
+        and dimension names should be if there is an index in ``value``.
     mutable : bool, optional
-        Switches between creating a :class:`~aesara.compile.sharedvalue.SharedVariable`
-        (``mutable=True``) vs. creating a :class:`~aesara.tensor.TensorConstant`
+        Switches between creating a :class:`~pytensor.compile.sharedvalue.SharedVariable`
+        (``mutable=True``) vs. creating a :class:`~pytensor.tensor.TensorConstant`
         (``mutable=False``).
         Consider using :class:`pymc.ConstantData` or :class:`pymc.MutableData` as less
         verbose alternatives to ``pm.Data(..., mutable=...)``.
@@ -601,7 +382,7 @@ def Data(
         version of the package. Since ``v4.1.0`` the default value is
         ``mutable=False``, with previous versions having ``mutable=True``.
     **kwargs : dict, optional
-        Extra arguments passed to :func:`aesara.shared`.
+        Extra arguments passed to :func:`pytensor.shared`.
 
     Examples
     --------
@@ -624,6 +405,9 @@ def Data(
     ...         model.set_data('data', data_vals)
     ...         idatas.append(pm.sample())
     """
+    if coords is None:
+        coords = {}
+
     if isinstance(value, list):
         value = np.array(value)
 
@@ -636,25 +420,28 @@ def Data(
         )
     name = model.name_for(name)
 
-    # `pandas_to_array` takes care of parameter `value` and
-    # transforms it to something digestible for Aesara.
-    arr = pandas_to_array(value)
+    # `convert_observed_data` takes care of parameter `value` and
+    # transforms it to something digestible for PyTensor.
+    arr = convert_observed_data(value)
+    if isinstance(arr, np.ma.MaskedArray):
+        raise NotImplementedError(
+            "Masked arrays or arrays with `nan` entries are not supported. "
+            "Pass them directly to `observed` if you want to trigger auto-imputation"
+        )
 
     if mutable is None:
-        major, minor = (int(v) for v in pm.__version__.split(".")[:2])
-        mutable = major == 4 and minor < 1
-        if mutable:
-            warnings.warn(
-                "The `mutable` kwarg was not specified. Currently it defaults to `pm.Data(mutable=True)`,"
-                " which is equivalent to using `pm.MutableData()`."
-                " In v4.1.0 the default will change to `pm.Data(mutable=False)`, equivalent to `pm.ConstantData`."
-                " Set `pm.Data(..., mutable=False/True)`, or use `pm.ConstantData`/`pm.MutableData`.",
-                FutureWarning,
-            )
+        warnings.warn(
+            "The `mutable` kwarg was not specified. Before v4.1.0 it defaulted to `pm.Data(mutable=True)`,"
+            " which is equivalent to using `pm.MutableData()`."
+            " In v4.1.0 the default changed to `pm.Data(mutable=False)`, equivalent to `pm.ConstantData`."
+            " Use `pm.ConstantData`/`pm.MutableData` or pass `pm.Data(..., mutable=False/True)` to avoid this warning.",
+            UserWarning,
+        )
+        mutable = False
     if mutable:
-        x = aesara.shared(arr, name, **kwargs)
+        x = pytensor.shared(arr, name, **kwargs)
     else:
-        x = at.as_tensor_variable(arr, name, **kwargs)
+        x = pt.as_tensor_variable(arr, name, **kwargs)
 
     if isinstance(dims, str):
         dims = (dims,)
@@ -665,16 +452,36 @@ def Data(
             expected=x.ndim,
         )
 
-    coords = determine_coords(model, value, dims)
-
+    # Optionally infer coords and dims from the input value.
     if export_index_as_coords:
-        model.add_coords(coords)
-    elif dims:
+        infer_dims_and_coords = export_index_as_coords
+        warnings.warn(
+            "Deprecation warning: 'export_index_as_coords; is deprecated and will be removed in future versions. Please use 'infer_dims_and_coords' instead.",
+            DeprecationWarning,
+        )
+
+    if infer_dims_and_coords:
+        coords, dims = determine_coords(model, value, dims)
+
+    if dims:
+        if not mutable:
+            # Use the dimension lengths from the before it was tensorified.
+            # These can still be tensors, but in many cases they are numeric.
+            xshape = np.shape(arr)
+        else:
+            xshape = x.shape
         # Register new dimension lengths
         for d, dname in enumerate(dims):
-            if not dname in model.dim_lengths:
-                model.add_coord(dname, values=None, length=x.shape[d])
+            if dname not in model.dim_lengths:
+                model.add_coord(
+                    name=dname,
+                    # Note: Coordinate values can't be taken from
+                    # the value, because it could be N-dimensional.
+                    values=coords.get(dname, None),
+                    mutable=mutable,
+                    length=xshape[d],
+                )
 
-    model.add_random_variable(x, dims=dims)
+    model.add_named_variable(x, dims=dims)
 
     return x
