@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -13,11 +13,13 @@
 #   limitations under the License.
 import warnings
 
-import aesara.tensor as at
 import numpy as np
+import pytensor.tensor as pt
 
-from aesara.tensor.random.basic import (
+from pytensor.tensor import TensorConstant
+from pytensor.tensor.random.basic import (
     RandomVariable,
+    ScipyRandomVariable,
     bernoulli,
     betabinom,
     binomial,
@@ -31,10 +33,11 @@ from scipy import stats
 
 import pymc as pm
 
-from pymc.aesaraf import floatX, intX, take_along_axis
 from pymc.distributions.dist_math import (
     betaln,
     binomln,
+    check_icdf_parameters,
+    check_icdf_value,
     check_parameters,
     factln,
     log_diff_normal_cdf,
@@ -43,11 +46,10 @@ from pymc.distributions.dist_math import (
     normal_lcdf,
 )
 from pymc.distributions.distribution import Discrete
-from pymc.distributions.logprob import logp
-from pymc.distributions.mixture import Mixture
 from pymc.distributions.shape_utils import rv_size_is_none
+from pymc.logprob.basic import logcdf, logp
 from pymc.math import sigmoid
-from pymc.vartypes import continuous_types
+from pymc.pytensorf import floatX, intX
 
 __all__ = [
     "Binomial",
@@ -56,10 +58,6 @@ __all__ = [
     "DiscreteWeibull",
     "Poisson",
     "NegativeBinomial",
-    "Constant",
-    "ZeroInflatedPoisson",
-    "ZeroInflatedBinomial",
-    "ZeroInflatedNegativeBinomial",
     "DiscreteUniform",
     "Geometric",
     "HyperGeometric",
@@ -81,6 +79,7 @@ class Binomial(Discrete):
     .. math:: f(x \mid n, p) = \binom{n}{x} p^x (1-p)^{n-x}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -106,78 +105,66 @@ class Binomial(Discrete):
 
     Parameters
     ----------
-    n: int
+    n : tensor_like of int
         Number of Bernoulli trials (n >= 0).
-    p: float
+    p : tensor_like of float
         Probability of success in each trial (0 < p < 1).
+    logit_p : tensor_like of float
+        Alternative log odds for the probability of success.
     """
     rv_op = binomial
 
     @classmethod
-    def dist(cls, n, p, *args, **kwargs):
-        n = at.as_tensor_variable(intX(n))
-        p = at.as_tensor_variable(floatX(p))
+    def dist(cls, n, p=None, logit_p=None, *args, **kwargs):
+        if p is not None and logit_p is not None:
+            raise ValueError("Incompatible parametrization. Can't specify both p and logit_p.")
+        elif p is None and logit_p is None:
+            raise ValueError("Incompatible parametrization. Must specify either p or logit_p.")
+
+        if logit_p is not None:
+            p = pt.sigmoid(logit_p)
+
+        n = pt.as_tensor_variable(intX(n))
+        p = pt.as_tensor_variable(floatX(p))
         return super().dist([n, p], **kwargs)
 
-    def get_moment(rv, size, n, p):
-        mean = at.round(n * p)
+    def moment(rv, size, n, p):
+        mean = pt.round(n * p)
         if not rv_size_is_none(size):
-            mean = at.full(size, mean)
+            mean = pt.full(size, mean)
         return mean
 
     def logp(value, n, p):
-        r"""
-        Calculate log-probability of Binomial distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.or_(at.lt(value, 0), at.gt(value, n)),
+        res = pt.switch(
+            pt.or_(pt.lt(value, 0), pt.gt(value, n)),
             -np.inf,
             binomln(n, value) + logpow(p, value) + logpow(1 - p, n - value),
         )
 
-        return check_parameters(res, 0 <= n, 0 <= p, p <= 1, msg="n >= 0, 0 <= p <= 1")
+        return check_parameters(
+            res,
+            n >= 0,
+            0 <= p,
+            p <= 1,
+            msg="n >= 0, 0 <= p <= 1",
+        )
 
     def logcdf(value, n, p):
-        """
-        Compute the log of the cumulative distribution function for Binomial distribution
-        at the specified value.
+        value = pt.floor(value)
 
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-        value = at.floor(value)
-
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.switch(
-                at.lt(value, n),
-                at.log(at.betainc(n - value, value + 1, 1 - p)),
+            pt.switch(
+                pt.lt(value, n),
+                pt.log(pt.betainc(n - value, value + 1, 1 - p)),
                 0,
             ),
         )
 
         return check_parameters(
             res,
-            0 <= n,
+            n >= 0,
             0 <= p,
             p <= 1,
             msg="n >= 0, 0 <= p <= 1",
@@ -199,6 +186,7 @@ class BetaBinomial(Discrete):
            \frac{B(x + \alpha, n - x + \beta)}{B(\alpha, \beta)}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -227,16 +215,16 @@ class BetaBinomial(Discrete):
     ========  =================================================================
     Support   :math:`x \in \{0, 1, \ldots, n\}`
     Mean      :math:`n \dfrac{\alpha}{\alpha + \beta}`
-    Variance  :math:`n \dfrac{\alpha \beta}{(\alpha+\beta)^2 (\alpha+\beta+1)}`
+    Variance  :math:`\dfrac{n \alpha \beta (\alpha+\beta+n)}{(\alpha+\beta)^2 (\alpha+\beta+1)}`
     ========  =================================================================
 
     Parameters
     ----------
-    n: int
+    n : tensor_like of int
         Number of Bernoulli trials (n >= 0).
-    alpha: float
+    alpha : tensor_like of float
         alpha > 0.
-    beta: float
+    beta : tensor_like of float
         beta > 0.
     """
 
@@ -244,75 +232,61 @@ class BetaBinomial(Discrete):
 
     @classmethod
     def dist(cls, alpha, beta, n, *args, **kwargs):
-        alpha = at.as_tensor_variable(floatX(alpha))
-        beta = at.as_tensor_variable(floatX(beta))
-        n = at.as_tensor_variable(intX(n))
+        alpha = pt.as_tensor_variable(floatX(alpha))
+        beta = pt.as_tensor_variable(floatX(beta))
+        n = pt.as_tensor_variable(intX(n))
         return super().dist([n, alpha, beta], **kwargs)
 
-    def get_moment(rv, size, n, alpha, beta):
-        mean = at.round((n * alpha) / (alpha + beta))
+    def moment(rv, size, n, alpha, beta):
+        mean = pt.round((n * alpha) / (alpha + beta))
         if not rv_size_is_none(size):
-            mean = at.full(size, mean)
+            mean = pt.full(size, mean)
         return mean
 
     def logp(value, n, alpha, beta):
-        r"""
-        Calculate log-probability of BetaBinomial distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-        res = at.switch(
-            at.or_(at.lt(value, 0), at.gt(value, n)),
+        res = pt.switch(
+            pt.or_(pt.lt(value, 0), pt.gt(value, n)),
             -np.inf,
             binomln(n, value) + betaln(value + alpha, n - value + beta) - betaln(alpha, beta),
         )
-        return check_parameters(res, n >= 0, alpha > 0, beta > 0, msg="n >= 0, alpha > 0, beta > 0")
+        return check_parameters(
+            res,
+            n >= 0,
+            alpha > 0,
+            beta > 0,
+            msg="n >= 0, alpha > 0, beta > 0",
+        )
 
     def logcdf(value, n, alpha, beta):
-        """
-        Compute the log of the cumulative distribution function for BetaBinomial distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value for which log CDF is calculated.
-
-        Returns
-        -------
-        TensorVariable
-        """
         # logcdf can only handle scalar values at the moment
         if np.ndim(value):
             raise TypeError(
                 f"BetaBinomial.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
             )
 
-        safe_lower = at.switch(at.lt(value, 0), value, 0)
-        res = at.switch(
-            at.lt(value, 0),
+        safe_lower = pt.switch(pt.lt(value, 0), value, 0)
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.switch(
-                at.lt(value, n),
-                at.logsumexp(
+            pt.switch(
+                pt.lt(value, n),
+                pt.logsumexp(
                     logp(
                         BetaBinomial.dist(alpha=alpha, beta=beta, n=n),
-                        at.arange(safe_lower, value + 1),
+                        pt.arange(safe_lower, value + 1),
                     ),
                     keepdims=False,
                 ),
                 0,
             ),
         )
-        return check_parameters(res, 0 <= n, 0 < alpha, 0 < beta, msg="n >= 0, alpha > 0, beta > 0")
+        return check_parameters(
+            res,
+            n >= 0,
+            alpha > 0,
+            beta > 0,
+            msg="n >= 0, alpha > 0, beta > 0",
+        )
 
 
 class Bernoulli(Discrete):
@@ -325,6 +299,7 @@ class Bernoulli(Discrete):
     .. math:: f(x \mid p) = p^{x} (1-p)^{1-x}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -354,9 +329,9 @@ class Bernoulli(Discrete):
 
     Parameters
     ----------
-    p: float
+    p : tensor_like of float
         Probability of success (0 < p < 1).
-    logit_p: float
+    logit_p : tensor_like of float
         Alternative log odds for the probability of success.
     """
     rv_op = bernoulli
@@ -369,64 +344,46 @@ class Bernoulli(Discrete):
             raise ValueError("Incompatible parametrization. Must specify either p or logit_p.")
 
         if logit_p is not None:
-            p = at.sigmoid(logit_p)
+            p = pt.sigmoid(logit_p)
 
-        p = at.as_tensor_variable(floatX(p))
+        p = pt.as_tensor_variable(floatX(p))
         return super().dist([p], **kwargs)
 
-    def get_moment(rv, size, p):
+    def moment(rv, size, p):
         if not rv_size_is_none(size):
-            p = at.full(size, p)
-        return at.switch(p < 0.5, 0, 1)
+            p = pt.full(size, p)
+        return pt.switch(p < 0.5, 0, 1)
 
     def logp(value, p):
-        r"""
-        Calculate log-probability of Bernoulli distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.or_(at.lt(value, 0), at.gt(value, 1)),
+        res = pt.switch(
+            pt.or_(pt.lt(value, 0), pt.gt(value, 1)),
             -np.inf,
-            at.switch(value, at.log(p), at.log1p(-p)),
+            pt.switch(value, pt.log(p), pt.log1p(-p)),
         )
 
-        return check_parameters(res, p >= 0, p <= 1, msg="0 <= p <= 1")
+        return check_parameters(
+            res,
+            0 <= p,
+            p <= 1,
+            msg="0 <= p <= 1",
+        )
 
     def logcdf(value, p):
-        """
-        Compute the log of the cumulative distribution function for Bernoulli distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.switch(
-                at.lt(value, 1),
-                at.log1p(-p),
+            pt.switch(
+                pt.lt(value, 1),
+                pt.log1p(-p),
                 0,
             ),
         )
-        return check_parameters(res, 0 <= p, p <= 1, msg="0 <= p <= 1")
+        return check_parameters(
+            res,
+            0 <= p,
+            p <= 1,
+            msg="0 <= p <= 1",
+        )
 
 
 class DiscreteWeibullRV(RandomVariable):
@@ -446,7 +403,7 @@ discrete_weibull = DiscreteWeibullRV()
 
 
 class DiscreteWeibull(Discrete):
-    R"""Discrete Weibull log-likelihood
+    R"""Discrete Weibull log-likelihood.
 
     The discrete Weibull distribution is a flexible model of count data that
     can handle both over- and under-dispersion.
@@ -455,6 +412,7 @@ class DiscreteWeibull(Discrete):
     .. math:: f(x \mid q, \beta) = q^{x^{\beta}} - q^{(x + 1)^{\beta}}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -483,66 +441,57 @@ class DiscreteWeibull(Discrete):
     Mean      :math:`\mu = \sum_{x = 1}^{\infty} q^{x^{\beta}}`
     Variance  :math:`2 \sum_{x = 1}^{\infty} x q^{x^{\beta}} - \mu - \mu^2`
     ========  ======================
+
+    Parameters
+    ----------
+    q : tensor_like of float
+        Shape parameter (0 < q < 1).
+    beta : tensor_like of float
+        Shape parameter (beta > 0).
+
     """
     rv_op = discrete_weibull
 
     @classmethod
     def dist(cls, q, beta, *args, **kwargs):
-        q = at.as_tensor_variable(floatX(q))
-        beta = at.as_tensor_variable(floatX(beta))
+        q = pt.as_tensor_variable(floatX(q))
+        beta = pt.as_tensor_variable(floatX(beta))
         return super().dist([q, beta], **kwargs)
 
-    def get_moment(rv, size, q, beta):
-        median = at.power(at.log(0.5) / at.log(q), 1 / beta) - 1
+    def moment(rv, size, q, beta):
+        median = pt.power(pt.log(0.5) / pt.log(q), 1 / beta) - 1
         if not rv_size_is_none(size):
-            median = at.full(size, median)
+            median = pt.full(size, median)
         return median
 
     def logp(value, q, beta):
-        r"""
-        Calculate log-probability of DiscreteWeibull distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.log(at.power(q, at.power(value, beta)) - at.power(q, at.power(value + 1, beta))),
+            pt.log(pt.power(q, pt.power(value, beta)) - pt.power(q, pt.power(value + 1, beta))),
         )
 
-        return check_parameters(res, 0 < q, q < 1, 0 < beta, msg="0 < q < 1, beta > 0")
+        return check_parameters(
+            res,
+            0 < q,
+            q < 1,
+            beta > 0,
+            msg="0 < q < 1, beta > 0",
+        )
 
     def logcdf(value, q, beta):
-        """
-        Compute the log of the cumulative distribution function for Discrete Weibull distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.log1p(-at.power(q, at.power(value + 1, beta))),
+            pt.log1p(-pt.power(q, pt.power(value + 1, beta))),
         )
-        return check_parameters(res, 0 < q, q < 1, 0 < beta, msg="0 < q < 1, beta > 0")
+        return check_parameters(
+            res,
+            0 < q,
+            q < 1,
+            beta > 0,
+            msg="0 < q < 1, beta > 0",
+        )
 
 
 class Poisson(Discrete):
@@ -556,6 +505,7 @@ class Poisson(Discrete):
     .. math:: f(x \mid \mu) = \frac{e^{-\mu}\mu^x}{x!}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -580,7 +530,7 @@ class Poisson(Discrete):
 
     Parameters
     ----------
-    mu: float
+    mu : tensor_like of float
         Expected number of occurrences during the given interval
         (mu >= 0).
 
@@ -593,65 +543,50 @@ class Poisson(Discrete):
 
     @classmethod
     def dist(cls, mu, *args, **kwargs):
-        mu = at.as_tensor_variable(floatX(mu))
+        mu = pt.as_tensor_variable(floatX(mu))
         return super().dist([mu], *args, **kwargs)
 
-    def get_moment(rv, size, mu):
-        mu = at.floor(mu)
+    def moment(rv, size, mu):
+        mu = pt.floor(mu)
         if not rv_size_is_none(size):
-            mu = at.full(size, mu)
+            mu = pt.full(size, mu)
         return mu
 
     def logp(value, mu):
-        r"""
-        Calculate log-probability of Poisson distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
             logpow(mu, value) - factln(value) - mu,
         )
-        log_prob = check_parameters(res, mu >= 0, msg="mu >= 0")
         # Return zero when mu and value are both zero
-        return at.switch(at.eq(mu, 0) * at.eq(value, 0), 0, log_prob)
-
-    def logcdf(value, mu):
-        """
-        Compute the log of the cumulative distribution function for Poisson distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-        value = at.floor(value)
-        # Avoid C-assertion when the gammaincc function is called with invalid values (#4340)
-        safe_mu = at.switch(at.lt(mu, 0), 0, mu)
-        safe_value = at.switch(at.lt(value, 0), 0, value)
-
-        res = at.switch(
-            at.lt(value, 0),
-            -np.inf,
-            at.log(at.gammaincc(safe_value + 1, safe_mu)),
+        res = pt.switch(
+            pt.eq(mu, 0) * pt.eq(value, 0),
+            0,
+            res,
+        )
+        return check_parameters(
+            res,
+            mu >= 0,
+            msg="mu >= 0",
         )
 
-        return check_parameters(res, 0 <= mu, msg="mu >= 0")
+    def logcdf(value, mu):
+        value = pt.floor(value)
+        # Avoid C-assertion when the gammaincc function is called with invalid values (#4340)
+        safe_mu = pt.switch(pt.lt(mu, 0), 0, mu)
+        safe_value = pt.switch(pt.lt(value, 0), 0, value)
+
+        res = pt.switch(
+            pt.lt(value, 0),
+            -np.inf,
+            pt.log(pt.gammaincc(safe_value + 1, safe_mu)),
+        )
+
+        return check_parameters(
+            res,
+            mu >= 0,
+            msg="mu >= 0",
+        )
 
 
 class NegativeBinomial(Discrete):
@@ -669,6 +604,7 @@ class NegativeBinomial(Discrete):
            (\alpha/(\mu+\alpha))^\alpha (\mu/(\mu+\alpha))^x
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -716,13 +652,13 @@ class NegativeBinomial(Discrete):
 
     Parameters
     ----------
-    alpha: float
+    alpha : tensor_like of float
         Gamma distribution shape parameter (alpha > 0).
-    mu: float
+    mu : tensor_like of float
         Gamma distribution mean (mu > 0).
-    p: float
+    p : tensor_like of float
         Alternative probability of success in each trial (0 < p < 1).
-    n: float
+    n : tensor_like of float
         Alternative number of target success trials (n > 0)
     """
     rv_op = nbinom
@@ -730,8 +666,8 @@ class NegativeBinomial(Discrete):
     @classmethod
     def dist(cls, mu=None, alpha=None, p=None, n=None, *args, **kwargs):
         n, p = cls.get_n_p(mu=mu, alpha=alpha, p=p, n=n)
-        n = at.as_tensor_variable(floatX(n))
-        p = at.as_tensor_variable(floatX(p))
+        n = pt.as_tensor_variable(floatX(n))
+        p = pt.as_tensor_variable(floatX(p))
         return super().dist([n, p], *args, **kwargs)
 
     @classmethod
@@ -754,31 +690,18 @@ class NegativeBinomial(Discrete):
 
         return n, p
 
-    def get_moment(rv, size, n, p):
-        mu = at.floor(n * (1 - p) / p)
+    def moment(rv, size, n, p):
+        mu = pt.floor(n * (1 - p) / p)
         if not rv_size_is_none(size):
-            mu = at.full(size, mu)
+            mu = pt.full(size, mu)
         return mu
 
     def logp(value, n, p):
-        r"""
-        Calculate log-probability of NegativeBinomial distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
         alpha = n
         mu = alpha * (1 - p) / p
 
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
             (
                 binomln(value + alpha - 1, value)
@@ -795,34 +718,20 @@ class NegativeBinomial(Discrete):
         )
 
         # Return Poisson when alpha gets very large.
-        return at.switch(at.gt(alpha, 1e10), logp(Poisson.dist(mu=mu), value), negbinom)
+        return pt.switch(pt.gt(alpha, 1e10), logp(Poisson.dist(mu=mu), value), negbinom)
 
     def logcdf(value, n, p):
-        """
-        Compute the log of the cumulative distribution function for NegativeBinomial distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.log(at.betainc(n, at.floor(value) + 1, p)),
+            pt.log(pt.betainc(n, pt.floor(value) + 1, p)),
         )
         return check_parameters(
             res,
-            0 < n,
+            n > 0,
             0 <= p,
             p <= 1,
-            msg="0 < n, 0 <= p <= 1",
+            msg="n > 0, 0 <= p <= 1",
         )
 
 
@@ -837,6 +746,7 @@ class Geometric(Discrete):
     .. math:: f(x \mid p) = p(1-p)^{x-1}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -860,7 +770,7 @@ class Geometric(Discrete):
 
     Parameters
     ----------
-    p: float
+    p : tensor_like of float
         Probability of success on an individual trial (0 < p <= 1).
     """
 
@@ -868,34 +778,20 @@ class Geometric(Discrete):
 
     @classmethod
     def dist(cls, p, *args, **kwargs):
-        p = at.as_tensor_variable(floatX(p))
+        p = pt.as_tensor_variable(floatX(p))
         return super().dist([p], *args, **kwargs)
 
-    def get_moment(rv, size, p):
-        mean = at.round(1.0 / p)
+    def moment(rv, size, p):
+        mean = pt.round(1.0 / p)
         if not rv_size_is_none(size):
-            mean = at.full(size, mean)
+            mean = pt.full(size, mean)
         return mean
 
     def logp(value, p):
-        r"""
-        Calculate log-probability of Geometric distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.lt(value, 1),
+        res = pt.switch(
+            pt.lt(value, 1),
             -np.inf,
-            at.log(p) + logpow(1 - p, value - 1),
+            pt.log(p) + logpow(1 - p, value - 1),
         )
 
         return check_parameters(
@@ -906,27 +802,26 @@ class Geometric(Discrete):
         )
 
     def logcdf(value, p):
-        """
-        Compute the log of the cumulative distribution function for Geometric distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.log1mexp(at.log1p(-p) * value),
+            pt.log1mexp(pt.log1p(-p) * value),
         )
         return check_parameters(
+            res,
+            0 <= p,
+            p <= 1,
+            msg="0 <= p <= 1",
+        )
+
+    def icdf(value, p):
+        res = pt.ceil(pt.log1p(-value) / pt.log1p(-p)).astype("int64")
+        res_1m = pt.maximum(res - 1, 0)
+        dist = pm.Geometric.dist(p=p)
+        value_1m = pt.exp(logcdf(dist, res_1m))
+        res = pt.switch(value_1m >= value, res_1m, res)
+        res = check_icdf_value(res, value)
+        return check_icdf_parameters(
             res,
             0 <= p,
             p <= 1,
@@ -946,6 +841,7 @@ class HyperGeometric(Discrete):
     .. math:: f(x \mid N, n, k) = \frac{\binom{k}{x}\binom{N-k}{n-x}}{\binom{N}{n}}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -957,9 +853,7 @@ class HyperGeometric(Discrete):
         k = 10
         for n in [20, 25]:
             pmf = st.hypergeom.pmf(x, N, k, n)
-            plt.plot(x, pmf, '-o', label='n = {}'.format(n))
-        plt.plot(x, pmf, '-o', label='N = {}'.format(N))
-        plt.plot(x, pmf, '-o', label='k = {}'.format(k))
+            plt.plot(x, pmf, '-o', label='N = {}, k = {}, n = {}'.format(N, k, n))
         plt.xlabel('x', fontsize=12)
         plt.ylabel('f(x)', fontsize=12)
         plt.legend(loc=1)
@@ -973,11 +867,11 @@ class HyperGeometric(Discrete):
 
     Parameters
     ----------
-    N : tensor_like of integer
+    N : tensor_like of int
         Total size of the population (N > 0)
-    k : tensor_like of integer
+    k : tensor_like of int
         Number of successful individuals in the population (0 <= k <= N)
-    n : tensor_like of integer
+    n : tensor_like of int
         Number of samples drawn from the population (0 <= n <= N)
     """
 
@@ -985,37 +879,19 @@ class HyperGeometric(Discrete):
 
     @classmethod
     def dist(cls, N, k, n, *args, **kwargs):
-        good = at.as_tensor_variable(intX(k))
-        bad = at.as_tensor_variable(intX(N - k))
-        n = at.as_tensor_variable(intX(n))
+        good = pt.as_tensor_variable(intX(k))
+        bad = pt.as_tensor_variable(intX(N - k))
+        n = pt.as_tensor_variable(intX(n))
         return super().dist([good, bad, n], *args, **kwargs)
 
-    def get_moment(rv, size, good, bad, n):
+    def moment(rv, size, good, bad, n):
         N, k = good + bad, good
-        mode = at.floor((n + 1) * (k + 1) / (N + 2))
+        mode = pt.floor((n + 1) * (k + 1) / (N + 2))
         if not rv_size_is_none(size):
-            mode = at.full(size, mode)
+            mode = pt.full(size, mode)
         return mode
 
     def logp(value, good, bad, n):
-        r"""
-        Calculate log-probability of HyperGeometric distribution at specified value.
-
-        Parameters
-        ----------
-        value : numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-        good : integer, array_like or TensorVariable
-            Number of successful individuals in the population. Alias for parameter :math:`k`.
-        bad : integer, array_like or TensorVariable
-            Number of unsuccessful individuals in the population. Alias for :math:`N-k`.
-
-        Returns
-        -------
-        TensorVariable
-        """
-
         tot = good + bad
         result = (
             betaln(good + 1, 1)
@@ -1026,41 +902,26 @@ class HyperGeometric(Discrete):
             - betaln(tot + 1, 1)
         )
         # value in [max(0, n - N + k), min(k, n)]
-        lower = at.switch(at.gt(n - tot + good, 0), n - tot + good, 0)
-        upper = at.switch(at.lt(good, n), good, n)
+        lower = pt.switch(pt.gt(n - tot + good, 0), n - tot + good, 0)
+        upper = pt.switch(pt.lt(good, n), good, n)
 
-        res = at.switch(
-            at.lt(value, lower),
+        res = pt.switch(
+            pt.lt(value, lower),
             -np.inf,
-            at.switch(
-                at.le(value, upper),
+            pt.switch(
+                pt.le(value, upper),
                 result,
                 -np.inf,
             ),
         )
 
-        return check_parameters(res, lower <= upper, msg="lower <= upper")
+        return check_parameters(
+            res,
+            lower <= upper,
+            msg="lower <= upper",
+        )
 
     def logcdf(value, good, bad, n):
-        """
-        Compute the log of the cumulative distribution function for HyperGeometric distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value : numeric
-            Value for which log CDF is calculated.
-        good : integer
-            Number of successful individuals in the population. Alias for parameter :math:`k`.
-        bad : integer
-            Number of unsuccessful individuals in the population. Alias for :math:`N-k`.
-        n : integer
-            Number of samples drawn from the population (0 <= n <= N)
-
-        Returns
-        -------
-        TensorVariable
-        """
         # logcdf can only handle scalar values at the moment
         if np.ndim(value):
             raise TypeError(
@@ -1069,15 +930,15 @@ class HyperGeometric(Discrete):
 
         N = good + bad
         # TODO: Use lower upper in locgdf for smarter logsumexp?
-        safe_lower = at.switch(at.lt(value, 0), value, 0)
+        safe_lower = pt.switch(pt.lt(value, 0), value, 0)
 
-        res = at.switch(
-            at.lt(value, 0),
+        res = pt.switch(
+            pt.lt(value, 0),
             -np.inf,
-            at.switch(
-                at.lt(value, n),
-                at.logsumexp(
-                    HyperGeometric.logp(at.arange(safe_lower, value + 1), good, bad, n),
+            pt.switch(
+                pt.lt(value, n),
+                pt.logsumexp(
+                    HyperGeometric.logp(pt.arange(safe_lower, value + 1), good, bad, n),
                     keepdims=False,
                 ),
                 0,
@@ -1086,16 +947,16 @@ class HyperGeometric(Discrete):
 
         return check_parameters(
             res,
-            0 < N,
+            N > 0,
             0 <= good,
-            0 <= n,
             good <= N,
+            0 <= n,
             n <= N,
             msg="N > 0, 0 <= good <= N, 0 <= n <= N",
         )
 
 
-class DiscreteUniformRV(RandomVariable):
+class DiscreteUniformRV(ScipyRandomVariable):
     name = "discrete_uniform"
     ndim_supp = 0
     ndims_params = [0, 0]
@@ -1103,7 +964,7 @@ class DiscreteUniformRV(RandomVariable):
     _print_name = ("DiscreteUniform", "\\operatorname{DiscreteUniform}")
 
     @classmethod
-    def rng_fn(cls, rng, lower, upper, size=None):
+    def rng_fn_scipy(cls, rng, lower, upper, size=None):
         return stats.randint.rvs(lower, upper + 1, size=size, random_state=rng)
 
 
@@ -1111,13 +972,14 @@ discrete_uniform = DiscreteUniformRV()
 
 
 class DiscreteUniform(Discrete):
-    R"""
-    Discrete uniform distribution.
+    R"""Discrete uniform distribution.
+
     The pmf of this distribution is
 
     .. math:: f(x \mid lower, upper) = \frac{1}{upper-lower+1}
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -1144,9 +1006,9 @@ class DiscreteUniform(Discrete):
 
     Parameters
     ----------
-    lower: int
+    lower : tensor_like of int
         Lower limit.
-    upper: int
+    upper : tensor_like of int
         Upper limit (upper > lower).
     """
 
@@ -1154,64 +1016,58 @@ class DiscreteUniform(Discrete):
 
     @classmethod
     def dist(cls, lower, upper, *args, **kwargs):
-        lower = intX(at.floor(lower))
-        upper = intX(at.floor(upper))
+        lower = intX(pt.floor(lower))
+        upper = intX(pt.floor(upper))
         return super().dist([lower, upper], **kwargs)
 
-    def get_moment(rv, size, lower, upper):
-        mode = at.maximum(at.floor((upper + lower) / 2.0), lower)
+    def moment(rv, size, lower, upper):
+        mode = pt.maximum(pt.floor((upper + lower) / 2.0), lower)
         if not rv_size_is_none(size):
-            mode = at.full(size, mode)
+            mode = pt.full(size, mode)
         return mode
 
     def logp(value, lower, upper):
-        r"""
-        Calculate log-probability of DiscreteUniform distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-        res = at.switch(
-            at.or_(at.lt(value, lower), at.gt(value, upper)),
+        res = pt.switch(
+            pt.or_(pt.lt(value, lower), pt.gt(value, upper)),
             -np.inf,
-            at.fill(value, -at.log(upper - lower + 1)),
+            pt.fill(value, -pt.log(upper - lower + 1)),
         )
-        return check_parameters(res, lower <= upper, msg="lower <= upper")
+        return check_parameters(
+            res,
+            lower <= upper,
+            msg="lower <= upper",
+        )
 
     def logcdf(value, lower, upper):
-        """
-        Compute the log of the cumulative distribution function for Discrete uniform distribution
-        at the specified value.
-
-        Parameters
-        ----------
-        value: numeric or np.ndarray or aesara.tensor
-            Value(s) for which log CDF is calculated. If the log CDF for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor.
-
-        Returns
-        -------
-        TensorVariable
-        """
-
-        res = at.switch(
-            at.le(value, lower),
+        res = pt.switch(
+            pt.le(value, lower),
             -np.inf,
-            at.switch(
-                at.lt(value, upper),
-                at.log(at.minimum(at.floor(value), upper) - lower + 1) - at.log(upper - lower + 1),
+            pt.switch(
+                pt.lt(value, upper),
+                pt.log(pt.minimum(pt.floor(value), upper) - lower + 1) - pt.log(upper - lower + 1),
                 0,
             ),
         )
 
-        return check_parameters(res, lower <= upper, msg="lower <= upper")
+        return check_parameters(
+            res,
+            lower <= upper,
+            msg="lower <= upper",
+        )
+
+    def icdf(value, lower, upper):
+        res = pt.ceil(value * (upper - lower + 1)).astype("int64") + lower - 1
+        res_1m = pt.maximum(res - 1, lower)
+        dist = pm.DiscreteUniform.dist(lower=lower, upper=upper)
+        value_1m = pt.exp(logcdf(dist, res_1m))
+        res = pt.switch(value_1m >= value, res_1m, res)
+
+        res = check_icdf_value(res, value)
+        return check_icdf_parameters(
+            res,
+            lower <= upper,
+            msg="lower <= upper",
+        )
 
 
 class Categorical(Discrete):
@@ -1223,6 +1079,7 @@ class Categorical(Discrete):
     .. math:: f(x \mid p) = p_x
 
     .. plot::
+        :context: close-figs
 
         import matplotlib.pyplot as plt
         import numpy as np
@@ -1245,388 +1102,78 @@ class Categorical(Discrete):
 
     Parameters
     ----------
-    p: array of floats
-        p > 0 and the elements of p must sum to 1. They will be automatically
-        rescaled otherwise.
+    p : array of floats
+        p > 0 and the elements of p must sum to 1.
+    logit_p : float
+        Alternative log odds for the probability of success.
     """
     rv_op = categorical
 
     @classmethod
-    def dist(cls, p, **kwargs):
-        if isinstance(p, np.ndarray) or isinstance(p, list):
-            if (np.asarray(p) < 0).any():
-                raise ValueError(f"Negative `p` parameters are not valid, got: {p}")
-            p_sum = np.sum([p], axis=-1)
-            if not np.all(np.isclose(p_sum, 1.0)):
+    def dist(cls, p=None, logit_p=None, **kwargs):
+        if p is not None and logit_p is not None:
+            raise ValueError("Incompatible parametrization. Can't specify both p and logit_p.")
+        elif p is None and logit_p is None:
+            raise ValueError("Incompatible parametrization. Must specify either p or logit_p.")
+
+        if logit_p is not None:
+            p = pm.math.softmax(logit_p, axis=-1)
+
+        p = pt.as_tensor_variable(p)
+        if isinstance(p, TensorConstant):
+            p_ = np.asarray(p.data)
+            if np.any(p_ < 0):
+                raise ValueError(f"Negative `p` parameters are not valid, got: {p_}")
+            p_sum_ = np.sum([p_], axis=-1)
+            if not np.all(np.isclose(p_sum_, 1.0)):
                 warnings.warn(
-                    f"`p` parameters sum to {p_sum}, instead of 1.0. They will be automatically rescaled. You can rescale them directly to get rid of this warning.",
+                    f"`p` parameters sum to {p_sum_}, instead of 1.0. "
+                    "They will be automatically rescaled. "
+                    "You can rescale them directly to get rid of this warning.",
                     UserWarning,
                 )
-                p = p / at.sum(p, axis=-1, keepdims=True)
-        p = at.as_tensor_variable(floatX(p))
+                p_ = p_ / pt.sum(p_, axis=-1, keepdims=True)
+                p = pt.as_tensor_variable(p_)
         return super().dist([p], **kwargs)
 
-    def get_moment(rv, size, p):
-        mode = at.argmax(p, axis=-1)
+    def moment(rv, size, p):
+        mode = pt.argmax(p, axis=-1)
         if not rv_size_is_none(size):
-            mode = at.full(size, mode)
+            mode = pt.full(size, mode)
         return mode
 
     def logp(value, p):
-        r"""
-        Calculate log-probability of Categorical distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or `TensorVariable`
-
-        """
-        k = at.shape(p)[-1]
+        k = pt.shape(p)[-1]
         p_ = p
-        value_clip = at.clip(value, 0, k - 1)
+        value_clip = pt.clip(value, 0, k - 1)
 
         if p.ndim > 1:
             if p.ndim > value_clip.ndim:
-                value_clip = at.shape_padleft(value_clip, p_.ndim - value_clip.ndim)
+                value_clip = pt.shape_padleft(value_clip, p_.ndim - value_clip.ndim)
             elif p.ndim < value_clip.ndim:
-                p = at.shape_padleft(p, value_clip.ndim - p_.ndim)
+                p = pt.shape_padleft(p, value_clip.ndim - p_.ndim)
             pattern = (p.ndim - 1,) + tuple(range(p.ndim - 1))
-            a = at.log(
-                take_along_axis(
+            a = pt.log(
+                pt.take_along_axis(
                     p.dimshuffle(pattern),
                     value_clip,
                 )
             )
         else:
-            a = at.log(p[value_clip])
+            a = pt.log(p[value_clip])
 
-        res = at.switch(
-            at.or_(at.lt(value, 0), at.gt(value, k - 1)),
+        res = pt.switch(
+            pt.or_(pt.lt(value, 0), pt.gt(value, k - 1)),
             -np.inf,
             a,
         )
 
         return check_parameters(
-            res, at.all(p_ >= 0, axis=-1), at.all(p <= 1, axis=-1), msg="0 <= p <=1"
-        )
-
-
-class ConstantRV(RandomVariable):
-    name = "constant"
-    ndim_supp = 0
-    ndims_params = [0]
-    _print_name = ("Constant", "\\operatorname{Constant}")
-
-    def make_node(self, rng, size, dtype, c):
-        c = at.as_tensor_variable(c)
-        return super().make_node(rng, size, c.dtype, c)
-
-    @classmethod
-    def rng_fn(cls, rng, c, size=None):
-        if size is None:
-            return c.copy()
-        return np.full(size, c)
-
-
-constant = ConstantRV()
-
-
-class Constant(Discrete):
-    r"""
-    Constant log-likelihood.
-
-    Parameters
-    ----------
-    c: float or int
-        Constant parameter. The dtype of `c` determines the dtype of the distribution.
-        This can affect which sampler is assigned to Constant variables, or variables
-        that use Constant, such as Mixtures.
-    """
-
-    rv_op = constant
-
-    @classmethod
-    def dist(cls, c, *args, **kwargs):
-        c = at.as_tensor_variable(c)
-        if c.dtype in continuous_types:
-            c = floatX(c)
-        return super().dist([c], **kwargs)
-
-    def get_moment(rv, size, c):
-        if not rv_size_is_none(size):
-            c = at.full(size, c)
-        return c
-
-    def logp(value, c):
-        r"""
-        Calculate log-probability of Constant distribution at specified value.
-
-        Parameters
-        ----------
-        value: numeric
-            Value(s) for which log-probability is calculated. If the log probabilities for multiple
-            values are desired the values must be provided in a numpy array or Aesara tensor
-
-        Returns
-        -------
-        TensorVariable
-        """
-        return at.switch(
-            at.eq(value, c),
-            at.zeros_like(value),
-            -np.inf,
-        )
-
-    def logcdf(value, c):
-        return at.switch(
-            at.lt(value, c),
-            -np.inf,
-            0,
-        )
-
-
-def _zero_inflated_mixture(*, name, nonzero_p, nonzero_dist, **kwargs):
-    """Helper function to create a zero-inflated mixture
-
-    If name is `None`, this function returns an unregistered variable
-    """
-    nonzero_p = at.as_tensor_variable(floatX(nonzero_p))
-    weights = at.stack([1 - nonzero_p, nonzero_p], axis=-1)
-    comp_dists = [
-        Constant.dist(0),
-        nonzero_dist,
-    ]
-    if name is not None:
-        return Mixture(name, weights, comp_dists, **kwargs)
-    else:
-        return Mixture.dist(weights, comp_dists, **kwargs)
-
-
-class ZeroInflatedPoisson:
-    R"""
-    Zero-inflated Poisson log-likelihood.
-
-    Often used to model the number of events occurring in a fixed period
-    of time when the times at which events occur are independent.
-    The pmf of this distribution is
-
-    .. math::
-
-        f(x \mid \psi, \mu) = \left\{ \begin{array}{l}
-            (1-\psi) + \psi e^{-\mu}, \text{if } x = 0 \\
-            \psi \frac{e^{-\theta}\theta^x}{x!}, \text{if } x=1,2,3,\ldots
-            \end{array} \right.
-
-    .. plot::
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-        x = np.arange(0, 22)
-        psis = [0.7, 0.4]
-        mus = [8, 4]
-        for psi, mu in zip(psis, mus):
-            pmf = st.poisson.pmf(x, mu)
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            plt.plot(x, pmf, '-o', label='$\\psi$ = {}, $\\mu$ = {}'.format(psi, mu))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi\mu`
-    Variance  :math:`\mu + \frac{1-\psi}{\psi}\mu^2`
-    ========  ==========================
-
-    Parameters
-    ----------
-    psi: float
-        Expected proportion of Poisson variates (0 < psi < 1)
-    mu: float
-        Expected number of occurrences during the given interval
-        (mu >= 0).
-    """
-
-    def __new__(cls, name, psi, mu, **kwargs):
-        return _zero_inflated_mixture(
-            name=name, nonzero_p=psi, nonzero_dist=Poisson.dist(mu=mu), **kwargs
-        )
-
-    @classmethod
-    def dist(cls, psi, mu, **kwargs):
-        return _zero_inflated_mixture(
-            name=None, nonzero_p=psi, nonzero_dist=Poisson.dist(mu=mu), **kwargs
-        )
-
-
-class ZeroInflatedBinomial:
-    R"""
-    Zero-inflated Binomial log-likelihood.
-
-    The pmf of this distribution is
-
-    .. math::
-
-        f(x \mid \psi, n, p) = \left\{ \begin{array}{l}
-            (1-\psi) + \psi (1-p)^{n}, \text{if } x = 0 \\
-            \psi {n \choose x} p^x (1-p)^{n-x}, \text{if } x=1,2,3,\ldots,n
-            \end{array} \right.
-
-    .. plot::
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-        x = np.arange(0, 25)
-        ns = [10, 20]
-        ps = [0.5, 0.7]
-        psis = [0.7, 0.4]
-        for n, p, psi in zip(ns, ps, psis):
-            pmf = st.binom.pmf(x, n, p)
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            plt.plot(x, pmf, '-o', label='n = {}, p = {}, $\\psi$ = {}'.format(n, p, psi))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi n p`
-    Variance  :math:`(1-\psi) n p [1 - p(1 - \psi n)].`
-    ========  ==========================
-
-    Parameters
-    ----------
-    psi: float
-        Expected proportion of Binomial variates (0 < psi < 1)
-    n: int
-        Number of Bernoulli trials (n >= 0).
-    p: float
-        Probability of success in each trial (0 < p < 1).
-
-    """
-
-    def __new__(cls, name, psi, n, p, **kwargs):
-        return _zero_inflated_mixture(
-            name=name, nonzero_p=psi, nonzero_dist=Binomial.dist(n=n, p=p), **kwargs
-        )
-
-    @classmethod
-    def dist(cls, psi, n, p, **kwargs):
-        return _zero_inflated_mixture(
-            name=None, nonzero_p=psi, nonzero_dist=Binomial.dist(n=n, p=p), **kwargs
-        )
-
-
-class ZeroInflatedNegativeBinomial:
-    R"""
-    Zero-Inflated Negative binomial log-likelihood.
-
-    The Zero-inflated version of the Negative Binomial (NB).
-    The NB distribution describes a Poisson random variable
-    whose rate parameter is gamma distributed.
-    The pmf of this distribution is
-
-    .. math::
-
-       f(x \mid \psi, \mu, \alpha) = \left\{
-         \begin{array}{l}
-           (1-\psi) + \psi \left (
-             \frac{\alpha}{\alpha+\mu}
-           \right) ^\alpha, \text{if } x = 0 \\
-           \psi \frac{\Gamma(x+\alpha)}{x! \Gamma(\alpha)} \left (
-             \frac{\alpha}{\mu+\alpha}
-           \right)^\alpha \left(
-             \frac{\mu}{\mu+\alpha}
-           \right)^x, \text{if } x=1,2,3,\ldots
-         \end{array}
-       \right.
-
-    .. plot::
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import scipy.stats as st
-        from scipy import special
-        import arviz as az
-        plt.style.use('arviz-darkgrid')
-
-        def ZeroInfNegBinom(a, m, psi, x):
-            pmf = special.binom(x + a - 1, x) * (a / (m + a))**a * (m / (m + a))**x
-            pmf[0] = (1 - psi) + pmf[0]
-            pmf[1:] =  psi * pmf[1:]
-            pmf /= pmf.sum()
-            return pmf
-
-        x = np.arange(0, 25)
-        alphas = [2, 4]
-        mus = [2, 8]
-        psis = [0.7, 0.7]
-        for a, m, psi in zip(alphas, mus, psis):
-            pmf = ZeroInfNegBinom(a, m, psi, x)
-            plt.plot(x, pmf, '-o', label=r'$\alpha$ = {}, $\mu$ = {}, $\psi$ = {}'.format(a, m, psi))
-        plt.xlabel('x', fontsize=12)
-        plt.ylabel('f(x)', fontsize=12)
-        plt.legend(loc=1)
-        plt.show()
-
-    ========  ==========================
-    Support   :math:`x \in \mathbb{N}_0`
-    Mean      :math:`\psi\mu`
-    Var       :math:`\psi\mu +  \left (1 + \frac{\mu}{\alpha} + \frac{1-\psi}{\mu} \right)`
-    ========  ==========================
-
-    The zero inflated negative binomial distribution can be parametrized
-    either in terms of mu or p, and either in terms of alpha or n.
-    The link between the parametrizations is given by
-
-    .. math::
-
-        \mu &= \frac{n(1-p)}{p} \\
-        \alpha &= n
-
-    Parameters
-    ----------
-    psi: float
-        Expected proportion of NegativeBinomial variates (0 < psi < 1)
-    mu: float
-        Poission distribution parameter (mu > 0).
-    alpha: float
-        Gamma distribution parameter (alpha > 0).
-    p: float
-        Alternative probability of success in each trial (0 < p < 1).
-    n: float
-        Alternative number of target success trials (n > 0)
-    """
-
-    def __new__(cls, name, psi, mu=None, alpha=None, p=None, n=None, **kwargs):
-        return _zero_inflated_mixture(
-            name=name,
-            nonzero_p=psi,
-            nonzero_dist=NegativeBinomial.dist(mu=mu, alpha=alpha, p=p, n=n),
-            **kwargs,
-        )
-
-    @classmethod
-    def dist(cls, psi, mu=None, alpha=None, p=None, n=None, **kwargs):
-        return _zero_inflated_mixture(
-            name=None,
-            nonzero_p=psi,
-            nonzero_dist=NegativeBinomial.dist(mu=mu, alpha=alpha, p=p, n=n),
-            **kwargs,
+            res,
+            0 <= p_,
+            p_ <= 1,
+            pt.isclose(pt.sum(p, axis=-1), 1),
+            msg="0 <= p <=1, sum(p) = 1",
         )
 
 
@@ -1639,15 +1186,15 @@ class _OrderedLogistic(Categorical):
 
     @classmethod
     def dist(cls, eta, cutpoints, *args, **kwargs):
-        eta = at.as_tensor_variable(floatX(eta))
-        cutpoints = at.as_tensor_variable(cutpoints)
+        eta = pt.as_tensor_variable(floatX(eta))
+        cutpoints = pt.as_tensor_variable(cutpoints)
 
-        pa = sigmoid(cutpoints - at.shape_padright(eta))
-        p_cum = at.concatenate(
+        pa = sigmoid(cutpoints - pt.shape_padright(eta))
+        p_cum = pt.concatenate(
             [
-                at.zeros_like(at.shape_padright(pa[..., 0])),
+                pt.zeros_like(pt.shape_padright(pa[..., 0])),
                 pa,
-                at.ones_like(at.shape_padright(pa[..., 0])),
+                pt.ones_like(pt.shape_padright(pa[..., 0])),
             ],
             axis=-1,
         )
@@ -1683,9 +1230,9 @@ class OrderedLogistic:
 
     Parameters
     ----------
-    eta: float
+    eta : tensor_like of float
         The predictor.
-    cutpoints: array
+    cutpoints : tensor_like of array
         The length K - 1 array of cutpoints which break :math:`\eta` into
         ranges. Do not explicitly set the first and last elements of
         :math:`c` to negative and positive infinity.
@@ -1696,7 +1243,6 @@ class OrderedLogistic:
 
     Examples
     --------
-
     .. code-block:: python
 
         # Generate data for a simple 1 dimensional example problem
@@ -1713,7 +1259,7 @@ class OrderedLogistic:
         # Ordered logistic regression
         with pm.Model() as model:
             cutpoints = pm.Normal("cutpoints", mu=[-1,1], sigma=10, shape=2,
-                                  transform=pm.distributions.transforms.ordered)
+                                  transform=pm.distributions.transforms.univariate_ordered)
             y_ = pm.OrderedLogistic("y", cutpoints=cutpoints, eta=x, observed=y)
             idata = pm.sample()
 
@@ -1746,22 +1292,22 @@ class _OrderedProbit(Categorical):
 
     @classmethod
     def dist(cls, eta, cutpoints, sigma=1, *args, **kwargs):
-        eta = at.as_tensor_variable(floatX(eta))
-        cutpoints = at.as_tensor_variable(cutpoints)
+        eta = pt.as_tensor_variable(floatX(eta))
+        cutpoints = pt.as_tensor_variable(cutpoints)
 
-        probits = at.shape_padright(eta) - cutpoints
-        _log_p = at.concatenate(
+        probits = pt.shape_padright(eta) - cutpoints
+        _log_p = pt.concatenate(
             [
-                at.shape_padright(normal_lccdf(0, sigma, probits[..., 0])),
+                pt.shape_padright(normal_lccdf(0, sigma, probits[..., 0])),
                 log_diff_normal_cdf(
-                    0, at.shape_padright(sigma), probits[..., :-1], probits[..., 1:]
+                    0, pt.shape_padright(sigma), probits[..., :-1], probits[..., 1:]
                 ),
-                at.shape_padright(normal_lcdf(0, sigma, probits[..., -1])),
+                pt.shape_padright(normal_lcdf(0, sigma, probits[..., -1])),
             ],
             axis=-1,
         )
-        _log_p = at.as_tensor_variable(floatX(_log_p))
-        p = at.exp(_log_p)
+        _log_p = pt.as_tensor_variable(floatX(_log_p))
+        p = pt.exp(_log_p)
 
         return super().dist(p, *args, **kwargs)
 
@@ -1796,19 +1342,20 @@ class OrderedProbit:
 
     Parameters
     ----------
-    eta: float
+    eta : tensor_like of float
         The predictor.
-    cutpoints: array
+    cutpoints : tensor_like array of floats
         The length K - 1 array of cutpoints which break :math:`\eta` into
         ranges. Do not explicitly set the first and last elements of
         :math:`c` to negative and positive infinity.
-    sigma: float, default 1.0
+    sigma : tensor_like of float, default 1.0
          Standard deviation of the probit function.
-    compute_p: boolean, default True
+    compute_p : boolean, default True
         Whether to compute and store in the trace the inferred probabilities of each categories,
         based on the cutpoints' values. Defaults to True.
         Might be useful to disable it if memory usage is of interest.
-    Example
+
+    Examples
     --------
     .. code:: python
 
