@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -74,3 +74,15 @@ class DtypeError(TypeError):
             super().__init__(f"{message} (expected {expected})")
         else:
             super().__init__(message)
+
+
+class TruncationError(RuntimeError):
+    """Exception for errors generated from truncated graphs"""
+
+
+class NotConstantValueError(ValueError):
+    pass
+
+
+class BlockModelAccessError(RuntimeError):
+    pass
