@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -22,3 +22,4 @@ from pymc.gp.gp import (
     MarginalKron,
     MarginalSparse,
 )
+from pymc.gp.hsgp_approx import HSGP
