@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -13,14 +13,18 @@
 #   limitations under the License.
 import warnings
 
-from collections import defaultdict, deque
-from typing import Dict, Iterator, NewType, Optional, Set
+from collections import defaultdict
+from typing import Dict, Iterable, List, NewType, Optional, Sequence, Set
 
-from aesara import function
-from aesara.compile.sharedvalue import SharedVariable
-from aesara.graph.basic import walk
-from aesara.tensor.random.op import RandomVariable
-from aesara.tensor.var import TensorConstant, TensorVariable
+from pytensor import function
+from pytensor.compile.sharedvalue import SharedVariable
+from pytensor.graph import Apply
+from pytensor.graph.basic import ancestors, walk
+from pytensor.scalar.basic import Cast
+from pytensor.tensor.elemwise import Elemwise
+from pytensor.tensor.random.op import RandomVariable
+from pytensor.tensor.shape import Shape
+from pytensor.tensor.var import TensorConstant, TensorVariable
 
 import pymc as pm
 
@@ -29,101 +33,126 @@ from pymc.util import get_default_varnames, get_var_name
 VarName = NewType("VarName", str)
 
 
+__all__ = (
+    "ModelGraph",
+    "model_to_graphviz",
+    "model_to_networkx",
+)
+
+
+def fast_eval(var):
+    return function([], var, mode="FAST_COMPILE")()
+
+
 class ModelGraph:
     def __init__(self, model):
         self.model = model
-        self.var_names = get_default_varnames(self.model.named_vars, include_transformed=False)
+        self._all_var_names = get_default_varnames(self.model.named_vars, include_transformed=False)
         self.var_list = self.model.named_vars.values()
-        self.transform_map = {
-            v.transformed: v.name for v in self.var_list if hasattr(v, "transformed")
+
+    def get_parent_names(self, var: TensorVariable) -> Set[VarName]:
+        if var.owner is None or var.owner.inputs is None:
+            return set()
+
+        def _filter_non_parameter_inputs(var):
+            node = var.owner
+            if isinstance(node.op, Shape):
+                # Don't show shape-related dependencies
+                return []
+            if isinstance(node.op, RandomVariable):
+                # Filter out rng, dtype and size parameters or RandomVariable nodes
+                return node.inputs[3:]
+            else:
+                # Otherwise return all inputs
+                return node.inputs
+
+        blockers = set(self.model.named_vars)
+
+        def _expand(x):
+            nonlocal blockers
+            if x.name in blockers:
+                return [x]
+            if isinstance(x.owner, Apply):
+                return reversed(_filter_non_parameter_inputs(x))
+            return []
+
+        parents = {
+            VarName(get_var_name(x))
+            for x in walk(nodes=_filter_non_parameter_inputs(var), expand=_expand)
+            # Only consider nodes that are in the named model variables.
+            if x.name and x.name in self._all_var_names
         }
-        self._deterministics = None
 
-    def get_deterministics(self, var):
-        """Compute the deterministic nodes of the graph, **not** including var itself."""
-        deterministics = []
-        attrs = ("transformed", "logpt")
-        for v in self.var_list:
-            if v != var and all(not hasattr(v, attr) for attr in attrs):
-                deterministics.append(v)
-        return deterministics
+        return parents
 
-    def _get_ancestors(self, var: TensorVariable, func) -> Set[TensorVariable]:
-        """Get all ancestors of a function, doing some accounting for deterministics."""
+    def vars_to_plot(self, var_names: Optional[Iterable[VarName]] = None) -> List[VarName]:
+        if var_names is None:
+            return self._all_var_names
 
-        # this contains all of the variables in the model EXCEPT var...
-        vars = set(self.var_list)
-        vars.remove(var)
+        selected_names = set(var_names)
 
-        blockers = set()  # type: Set[TensorVariable]
-        retval = set()  # type: Set[TensorVariable]
+        # .copy() because sets cannot change in size during iteration
+        for var_name in selected_names.copy():
+            if var_name not in self._all_var_names:
+                raise ValueError(f"{var_name} is not in this model.")
 
-        def _expand(node) -> Optional[Iterator[TensorVariable]]:
-            if node in blockers:
-                return None
-            elif node in vars:
-                blockers.add(node)
-                retval.add(node)
-                return None
-            elif node.owner:
-                blockers.add(node)
-                return reversed(node.owner.inputs)
-            else:
-                return None
+            for model_var in self.var_list:
+                if model_var in self.model.observed_RVs:
+                    if self.model.rvs_to_values[model_var] == self.model[var_name]:
+                        selected_names.add(model_var.name)
 
-        list(walk(deque([func]), _expand, bfs=True))
-        return retval
+        selected_ancestors = set(
+            filter(
+                lambda rv: rv.name in self._all_var_names,
+                list(ancestors([self.model[var_name] for var_name in selected_names])),
+            )
+        )
 
-    def _filter_parents(self, var, parents) -> Set[VarName]:
-        """Get direct parents of a var, as strings"""
-        keep = set()  # type: Set[VarName]
-        for p in parents:
-            if p == var:
-                continue
-            elif p.name in self.var_names:
-                keep.add(p.name)
-            elif p in self.transform_map:
-                if self.transform_map[p] != var.name:
-                    keep.add(self.transform_map[p])
-            else:
-                raise AssertionError(f"Do not know what to do with {get_var_name(p)}")
-        return keep
+        for var in selected_ancestors.copy():
+            if var in self.model.observed_RVs:
+                selected_ancestors.add(self.model.rvs_to_values[var])
 
-    def get_parents(self, var: TensorVariable) -> Set[VarName]:
-        """Get the named nodes that are direct inputs to the var"""
-        # TODO: Update these lines, variables no longer have a `logpt` attribute
-        if hasattr(var, "transformed"):
-            func = var.transformed.logpt
-        elif hasattr(var, "logpt"):
-            func = var.logpt
-        else:
-            func = var
+        # ordering of self._all_var_names is important
+        return [VarName(var.name) for var in selected_ancestors]
 
-        parents = self._get_ancestors(var, func)
-        return self._filter_parents(var, parents)
-
-    def make_compute_graph(self) -> Dict[str, Set[VarName]]:
+    def make_compute_graph(
+        self, var_names: Optional[Iterable[VarName]] = None
+    ) -> Dict[VarName, Set[VarName]]:
         """Get map of var_name -> set(input var names) for the model"""
-        input_map = defaultdict(set)  # type: Dict[str, Set[VarName]]
+        input_map: Dict[VarName, Set[VarName]] = defaultdict(set)
 
-        for var_name in self.var_names:
+        for var_name in self.vars_to_plot(var_names):
             var = self.model[var_name]
-            key = var_name
-            val = self.get_parents(var)
-            input_map[key] = input_map[key].union(val)
+            parent_name = self.get_parent_names(var)
+            input_map[var_name] = input_map[var_name].union(parent_name)
 
-            if hasattr(var.tag, "observations"):
-                try:
-                    obs_name = var.tag.observations.name
-                    if obs_name:
+            if var in self.model.observed_RVs:
+                obs_node = self.model.rvs_to_values[var]
+
+                # loop created so that the elif block can go through this again
+                # and remove any intermediate ops, notably dtype casting, to observations
+                while True:
+                    obs_name = obs_node.name
+                    if obs_name and obs_name != var_name:
                         input_map[var_name] = input_map[var_name].difference({obs_name})
                         input_map[obs_name] = input_map[obs_name].union({var_name})
-                except AttributeError:
-                    pass
+                        break
+                    elif (
+                        # for cases where observations are cast to a certain dtype
+                        # see issue 5795: https://github.com/pymc-devs/pymc/issues/5795
+                        obs_node.owner
+                        and isinstance(obs_node.owner.op, Elemwise)
+                        and isinstance(obs_node.owner.op.scalar_op, Cast)
+                    ):
+                        # we can retrieve the observation node by going up the graph
+                        obs_node = obs_node.owner.inputs[0]
+                    else:
+                        break
+
         return input_map
 
-    def _make_node(self, var_name, graph, *, formatting: str = "plain"):
-        """Attaches the given variable to a graphviz Digraph"""
+    def _make_node(self, var_name, graph, *, nx=False, cluster=False, formatting: str = "plain"):
+        """Attaches the given variable to a graphviz or networkx Digraph"""
         v = self.model[var_name]
 
         shape = None
@@ -142,15 +171,15 @@ class ModelGraph:
             shape = "box"
             style = "rounded, filled"
             label = f"{var_name}\n~\nMutableData"
-        elif v.owner and isinstance(v.owner.op, RandomVariable):
+        elif v in self.model.basic_RVs:
             shape = "ellipse"
-            if hasattr(v.tag, "observations"):
-                # observed RV
+            if v in self.model.observed_RVs:
                 style = "filled"
             else:
-                shape = "ellipse"
-                syle = None
-            symbol = v.owner.op.__class__.__name__.strip("RV")
+                style = None
+            symbol = v.owner.op.__class__.__name__
+            if symbol.endswith("RV"):
+                symbol = symbol[:-2]
             label = f"{var_name}\n~\n{symbol}"
         else:
             shape = "box"
@@ -163,12 +192,15 @@ class ModelGraph:
             "label": label,
         }
 
-        graph.node(var_name.replace(":", "&"), **kwargs)
+        if cluster:
+            kwargs["cluster"] = cluster
 
-    def _eval(self, var):
-        return function([], var, mode="FAST_COMPILE")()
+        if nx:
+            graph.add_node(var_name.replace(":", "&"), **kwargs)
+        else:
+            graph.node(var_name.replace(":", "&"), **kwargs)
 
-    def get_plates(self):
+    def get_plates(self, var_names: Optional[Iterable[VarName]] = None) -> Dict[str, Set[VarName]]:
         """Rough but surprisingly accurate plate detection.
 
         Just groups by the shape of the underlying distribution.  Will be wrong
@@ -176,22 +208,39 @@ class ModelGraph:
 
         Returns
         -------
-        dict: str -> set[str]
+        dict
+            Maps plate labels to the set of ``VarName``s inside the plate.
         """
         plates = defaultdict(set)
-        for var_name in self.var_names:
-            v = self.model[var_name]
-            if var_name in self.model.RV_dims:
-                plate_label = " x ".join(
-                    f"{d} ({self._eval(self.model.dim_lengths[d])})"
-                    for d in self.model.RV_dims[var_name]
-                )
-            else:
-                plate_label = " x ".join(map(str, self._eval(v.shape)))
-            plates[plate_label].add(var_name)
-        return plates
 
-    def make_graph(self, formatting: str = "plain"):
+        # TODO: Evaluate all RV shapes and dim_length at once.
+        #       This should help to find discrepancies, and
+        #       avoids unnecessary function compiles for deetermining labels.
+
+        for var_name in self.vars_to_plot(var_names):
+            v = self.model[var_name]
+            shape: Sequence[int] = fast_eval(v.shape)
+            dim_labels = []
+            if var_name in self.model.named_vars_to_dims:
+                # The RV is associated with `dims` information.
+                for d, dname in enumerate(self.model.named_vars_to_dims[var_name]):
+                    if dname is None:
+                        # Unnamed dimension in a `dims` tuple!
+                        dlen = shape[d]
+                        dname = f"{var_name}_dim{d}"
+                    else:
+                        dlen = fast_eval(self.model.dim_lengths[dname])
+                    dim_labels.append(f"{dname} ({dlen})")
+                plate_label = " x ".join(dim_labels)
+            else:
+                # The RV has no `dims` information.
+                dim_labels = [str(x) for x in shape]
+                plate_label = " x ".join(map(str, shape))
+            plates[plate_label].add(var_name)
+
+        return dict(plates)
+
+    def make_graph(self, var_names: Optional[Iterable[VarName]] = None, formatting: str = "plain"):
         """Make graphviz Digraph of PyMC model
 
         Returns
@@ -207,25 +256,152 @@ class ModelGraph:
                 "\tconda install -c conda-forge python-graphviz"
             )
         graph = graphviz.Digraph(self.model.name)
-        for plate_label, var_names in self.get_plates().items():
+        for plate_label, all_var_names in self.get_plates(var_names).items():
             if plate_label:
                 # must be preceded by 'cluster' to get a box around it
                 with graph.subgraph(name="cluster" + plate_label) as sub:
-                    for var_name in var_names:
+                    for var_name in all_var_names:
                         self._make_node(var_name, sub, formatting=formatting)
                     # plate label goes bottom right
                     sub.attr(label=plate_label, labeljust="r", labelloc="b", style="rounded")
             else:
-                for var_name in var_names:
+                for var_name in all_var_names:
                     self._make_node(var_name, graph, formatting=formatting)
 
-        for key, values in self.make_compute_graph().items():
-            for value in values:
-                graph.edge(value.replace(":", "&"), key.replace(":", "&"))
+        for child, parents in self.make_compute_graph(var_names=var_names).items():
+            # parents is a set of rv names that precede child rv nodes
+            for parent in parents:
+                graph.edge(parent.replace(":", "&"), child.replace(":", "&"))
+
         return graph
 
+    def make_networkx(
+        self, var_names: Optional[Iterable[VarName]] = None, formatting: str = "plain"
+    ):
+        """Make networkx Digraph of PyMC model
 
-def model_to_graphviz(model=None, *, formatting: str = "plain"):
+        Returns
+        -------
+        networkx.Digraph
+        """
+        try:
+            import networkx
+        except ImportError:
+            raise ImportError(
+                "This function requires the python library networkx, along with binaries. "
+                "The easiest way to install all of this is by running\n\n"
+                "\tconda install networkx"
+            )
+        graphnetwork = networkx.DiGraph(name=self.model.name)
+        for plate_label, all_var_names in self.get_plates(var_names).items():
+            if plate_label:
+                # # must be preceded by 'cluster' to get a box around it
+
+                subgraphnetwork = networkx.DiGraph(name="cluster" + plate_label, label=plate_label)
+
+                for var_name in all_var_names:
+                    self._make_node(
+                        var_name,
+                        subgraphnetwork,
+                        nx=True,
+                        cluster="cluster" + plate_label,
+                        formatting=formatting,
+                    )
+                for sgn in subgraphnetwork.nodes:
+                    networkx.set_node_attributes(
+                        subgraphnetwork,
+                        {sgn: {"labeljust": "r", "labelloc": "b", "style": "rounded"}},
+                    )
+                node_data = {
+                    e[0]: e[1]
+                    for e in graphnetwork.nodes(data=True) & subgraphnetwork.nodes(data=True)
+                }
+
+                graphnetwork = networkx.compose(graphnetwork, subgraphnetwork)
+                networkx.set_node_attributes(graphnetwork, node_data)
+                graphnetwork.graph["name"] = self.model.name
+            else:
+                for var_name in all_var_names:
+                    self._make_node(var_name, graphnetwork, nx=True, formatting=formatting)
+
+        for child, parents in self.make_compute_graph(var_names=var_names).items():
+            # parents is a set of rv names that precede child rv nodes
+            for parent in parents:
+                graphnetwork.add_edge(parent.replace(":", "&"), child.replace(":", "&"))
+        return graphnetwork
+
+
+def model_to_networkx(
+    model=None,
+    *,
+    var_names: Optional[Iterable[VarName]] = None,
+    formatting: str = "plain",
+):
+    """Produce a networkx Digraph from a PyMC model.
+
+    Requires networkx, which may be installed most easily with::
+
+        conda install networkx
+
+    Alternatively, you may install using pip with::
+
+        pip install networkx
+
+    See https://networkx.org/documentation/stable/ for more information.
+
+    Parameters
+    ----------
+    model : Model
+        The model to plot. Not required when called from inside a modelcontext.
+    var_names : iterable of str, optional
+        Subset of variables to be plotted that identify a subgraph with respect to the entire model graph
+    formatting : str, optional
+        one of { "plain" }
+
+    Examples
+    --------
+    How to plot the graph of the model.
+
+    .. code-block:: python
+
+        import numpy as np
+        from pymc import HalfCauchy, Model, Normal, model_to_networkx
+
+        J = 8
+        y = np.array([28, 8, -3, 7, -1, 1, 18, 12])
+        sigma = np.array([15, 10, 16, 11, 9, 11, 10, 18])
+
+        with Model() as schools:
+
+            eta = Normal("eta", 0, 1, shape=J)
+            mu = Normal("mu", 0, sigma=1e6)
+            tau = HalfCauchy("tau", 25)
+
+            theta = mu + tau * eta
+
+            obs = Normal("obs", theta, sigma=sigma, observed=y)
+
+        model_to_networkx(schools)
+    """
+    if "plain" not in formatting:
+        raise ValueError(f"Unsupported formatting for graph nodes: '{formatting}'. See docstring.")
+
+    if formatting != "plain":
+        warnings.warn(
+            "Formattings other than 'plain' are currently not supported.",
+            UserWarning,
+            stacklevel=2,
+        )
+    model = pm.modelcontext(model)
+    return ModelGraph(model).make_networkx(var_names=var_names, formatting=formatting)
+
+
+def model_to_graphviz(
+    model=None,
+    *,
+    var_names: Optional[Iterable[VarName]] = None,
+    formatting: str = "plain",
+):
     """Produce a graphviz Digraph from a PyMC model.
 
     Requires graphviz, which may be installed most easily with
@@ -240,7 +416,9 @@ def model_to_graphviz(model=None, *, formatting: str = "plain"):
     ----------
     model : pm.Model
         The model to plot. Not required when called from inside a modelcontext.
-    formatting : str
+    var_names : iterable of variable names, optional
+        Subset of variables to be plotted that identify a subgraph with respect to the entire model graph
+    formatting : str, optional
         one of { "plain" }
 
     Examples
@@ -268,11 +446,13 @@ def model_to_graphviz(model=None, *, formatting: str = "plain"):
 
         model_to_graphviz(schools)
     """
-    if not "plain" in formatting:
+    if "plain" not in formatting:
         raise ValueError(f"Unsupported formatting for graph nodes: '{formatting}'. See docstring.")
     if formatting != "plain":
         warnings.warn(
-            "Formattings other than 'plain' are currently not supported.", UserWarning, stacklevel=2
+            "Formattings other than 'plain' are currently not supported.",
+            UserWarning,
+            stacklevel=2,
         )
     model = pm.modelcontext(model)
-    return ModelGraph(model).make_graph(formatting=formatting)
+    return ModelGraph(model).make_graph(var_names=var_names, formatting=formatting)
