@@ -1,4 +1,4 @@
-#   Copyright 2021 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -16,11 +16,15 @@ import itertools
 
 from typing import Union
 
-from aesara.graph.basic import walk
-from aesara.tensor.basic import TensorVariable, Variable
-from aesara.tensor.elemwise import DimShuffle
-from aesara.tensor.random.basic import RandomVariable
-from aesara.tensor.var import TensorConstant
+from pytensor.compile import SharedVariable
+from pytensor.graph.basic import Constant, walk
+from pytensor.tensor.basic import TensorVariable, Variable
+from pytensor.tensor.elemwise import DimShuffle
+from pytensor.tensor.random.basic import RandomVariable
+from pytensor.tensor.random.var import (
+    RandomGeneratorSharedVariable,
+    RandomStateSharedVariable,
+)
 
 from pymc.model import Model
 
@@ -31,28 +35,64 @@ __all__ = [
 ]
 
 
-def str_for_dist(rv: TensorVariable, formatting: str = "plain", include_params: bool = True) -> str:
-    """Make a human-readable string representation of a RandomVariable in a model, either
+def str_for_dist(
+    dist: TensorVariable, formatting: str = "plain", include_params: bool = True
+) -> str:
+    """Make a human-readable string representation of a Distribution in a model, either
     LaTeX or plain, optionally with distribution parameter values included."""
 
     if include_params:
         # first 3 args are always (rng, size, dtype), rest is relevant for distribution
-        dist_args = [_str_for_input_var(x, formatting=formatting) for x in rv.owner.inputs[3:]]
+        if isinstance(dist.owner.op, RandomVariable):
+            dist_args = [
+                _str_for_input_var(x, formatting=formatting) for x in dist.owner.inputs[3:]
+            ]
+        else:
+            dist_args = [
+                _str_for_input_var(x, formatting=formatting)
+                for x in dist.owner.inputs
+                if not isinstance(x, (RandomStateSharedVariable, RandomGeneratorSharedVariable))
+            ]
 
-    print_name = rv.name if rv.name is not None else "<unnamed>"
+    print_name = dist.name
+
     if "latex" in formatting:
-        print_name = r"\text{" + _latex_escape(print_name) + "}"
-        dist_name = rv.owner.op._print_name[1]
+        if print_name is not None:
+            print_name = r"\text{" + _latex_escape(dist.name.strip("$")) + "}"
+
+        op_name = (
+            dist.owner.op._print_name[1]
+            if hasattr(dist.owner.op, "_print_name")
+            else r"\\operatorname{Unknown}"
+        )
         if include_params:
-            return r"${} \sim {}({})$".format(print_name, dist_name, ",~".join(dist_args))
+            if print_name:
+                return r"${} \sim {}({})$".format(
+                    print_name, op_name, ",~".join([d.strip("$") for d in dist_args])
+                )
+            else:
+                return r"${}({})$".format(op_name, ",~".join([d.strip("$") for d in dist_args]))
+
         else:
-            return rf"${print_name} \sim {dist_name}$"
+            if print_name:
+                return rf"${print_name} \sim {op_name}$"
+            else:
+                return rf"${op_name}$"
+
     else:  # plain
-        dist_name = rv.owner.op._print_name[0]
+        dist_name = (
+            dist.owner.op._print_name[0] if hasattr(dist.owner.op, "_print_name") else "Unknown"
+        )
         if include_params:
-            return r"{} ~ {}({})".format(print_name, dist_name, ", ".join(dist_args))
+            if print_name:
+                return r"{} ~ {}({})".format(print_name, dist_name, ", ".join(dist_args))
+            else:
+                return r"{}({})".format(dist_name, ", ".join(dist_args))
         else:
-            return rf"{print_name} ~ {dist_name}"
+            if print_name:
+                return rf"{print_name} ~ {dist_name}"
+            else:
+                return dist_name
 
 
 def str_for_model(model: Model, formatting: str = "plain", include_params: bool = True) -> str:
@@ -100,7 +140,7 @@ def str_for_potential_or_deterministic(
     LaTeX or plain, optionally with distribution parameter values included."""
     print_name = var.name if var.name is not None else "<unnamed>"
     if "latex" in formatting:
-        print_name = r"\text{" + _latex_escape(print_name) + "}"
+        print_name = r"\text{" + _latex_escape(print_name.strip("$")) + "}"
         if include_params:
             return rf"${print_name} \sim \operatorname{{{dist_name}}}({_str_for_expression(var, formatting=formatting)})$"
         else:
@@ -113,16 +153,21 @@ def str_for_potential_or_deterministic(
 
 
 def _str_for_input_var(var: Variable, formatting: str) -> str:
-    def _is_potential_or_determinstic(var: Variable) -> bool:
+    # Avoid circular import
+    from pymc.distributions.distribution import SymbolicRandomVariable
+
+    def _is_potential_or_deterministic(var: Variable) -> bool:
         try:
             return var.str_repr.__func__.func is str_for_potential_or_deterministic
         except AttributeError:
             # in case other code overrides str_repr, fallback
             return False
 
-    if isinstance(var, TensorConstant):
+    if isinstance(var, (Constant, SharedVariable)):
         return _str_for_constant(var, formatting)
-    elif isinstance(var.owner.op, RandomVariable) or _is_potential_or_determinstic(var):
+    elif isinstance(
+        var.owner.op, (RandomVariable, SymbolicRandomVariable)
+    ) or _is_potential_or_deterministic(var):
         # show the names for RandomVariables, Deterministics, and Potentials, rather
         # than the full expression
         return _str_for_input_rv(var, formatting)
@@ -133,41 +178,66 @@ def _str_for_input_var(var: Variable, formatting: str) -> str:
 
 
 def _str_for_input_rv(var: Variable, formatting: str) -> str:
-    _str = var.name if var.name is not None else "<unnamed>"
+    _str = (
+        var.name
+        if var.name is not None
+        else str_for_dist(var, formatting=formatting, include_params=True)
+    )
     if "latex" in formatting:
-        return r"\text{" + _latex_escape(_str) + "}"
+        return _latex_text_format(_latex_escape(_str.strip("$")))
     else:
         return _str
 
 
-def _str_for_constant(var: TensorConstant, formatting: str) -> str:
-    if len(var.data.shape) == 0:
-        return f"{var.data:.3g}"
-    elif len(var.data.shape) == 1 and var.data.shape[0] == 1:
-        return f"{var.data[0]:.3g}"
-    elif "latex" in formatting:
-        return r"\text{<constant>}"
+def _str_for_constant(var: Union[Constant, SharedVariable], formatting: str) -> str:
+    if isinstance(var, Constant):
+        var_data = var.data
+        var_type = "constant"
     else:
-        return r"<constant>"
+        var_data = var.get_value()
+        var_type = "shared"
+
+    if len(var_data.shape) == 0:
+        return f"{var_data:.3g}"
+    elif len(var_data.shape) == 1 and var_data.shape[0] == 1:
+        return f"{var_data[0]:.3g}"
+    elif "latex" in formatting:
+        return rf"\text{{<{var_type}>}}"
+    else:
+        return rf"<{var_type}>"
 
 
 def _str_for_expression(var: Variable, formatting: str) -> str:
+    # Avoid circular import
+    from pymc.distributions.distribution import SymbolicRandomVariable
+
     # construct a string like f(a1, ..., aN) listing all random variables a as arguments
     def _expand(x):
-        if x.owner and (not isinstance(x.owner.op, RandomVariable)):
+        if x.owner and (not isinstance(x.owner.op, (RandomVariable, SymbolicRandomVariable))):
             return reversed(x.owner.inputs)
 
     parents = [
         x
         for x in walk(nodes=var.owner.inputs, expand=_expand)
-        if x.owner and isinstance(x.owner.op, RandomVariable)
+        if x.owner and isinstance(x.owner.op, (RandomVariable, SymbolicRandomVariable))
     ]
     names = [x.name for x in parents]
 
     if "latex" in formatting:
-        return r"f(" + ",~".join([r"\text{" + _latex_escape(n) + "}" for n in names]) + ")"
+        return (
+            r"f("
+            + ",~".join([_latex_text_format(_latex_escape(n.strip("$"))) for n in names])
+            + ")"
+        )
     else:
-        return r"f(" + ", ".join(names) + ")"
+        return r"f(" + ", ".join([n.strip("$") for n in names]) + ")"
+
+
+def _latex_text_format(text: str) -> str:
+    if r"\operatorname{" in text:
+        return text
+    else:
+        return r"\text{" + text + "}"
 
 
 def _latex_escape(text: str) -> str:
