@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -11,27 +11,36 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Callable, Dict, List, Optional, Tuple
 
 import numpy as np
 import numpy.random as nr
+import pytensor
 import scipy.linalg
 import scipy.special
 
-from aesara.graph.fg import MissingInputError
-from aesara.tensor.random.basic import BernoulliRV, CategoricalRV
+from pytensor import tensor as pt
+from pytensor.graph.fg import MissingInputError
+from pytensor.tensor.random.basic import BernoulliRV, CategoricalRV
 
 import pymc as pm
 
-from pymc.aesaraf import compile_pymc, floatX, rvs_to_value_vars
 from pymc.blocking import DictToArrayBijection, RaveledVars
+from pymc.pytensorf import (
+    CallableTensor,
+    compile_pymc,
+    floatX,
+    join_nonshared_inputs,
+    replace_rng_nodes,
+)
 from pymc.step_methods.arraystep import (
     ArrayStep,
     ArrayStepShared,
-    Competence,
     PopulationArrayStepShared,
+    StatsType,
     metrop_select,
 )
+from pymc.step_methods.compound import Competence
 
 __all__ = [
     "Metropolis",
@@ -47,6 +56,8 @@ __all__ = [
     "MultivariateNormalProposal",
 ]
 
+from pymc.util import get_value_vars_from_user_vars
+
 # Available proposal distributions for Metropolis
 
 
@@ -57,38 +68,29 @@ class Proposal:
 
 class NormalProposal(Proposal):
     def __call__(self, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
-        return rng.normal(scale=self.s)
+        return (rng or nr).normal(scale=self.s)
 
 
 class UniformProposal(Proposal):
     def __call__(self, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
-        return rng.uniform(low=-self.s, high=self.s, size=len(self.s))
+        return (rng or nr).uniform(low=-self.s, high=self.s, size=len(self.s))
 
 
 class CauchyProposal(Proposal):
     def __call__(self, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
-        return rng.standard_cauchy(size=np.size(self.s)) * self.s
+        return (rng or nr).standard_cauchy(size=np.size(self.s)) * self.s
 
 
 class LaplaceProposal(Proposal):
     def __call__(self, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
         size = np.size(self.s)
-        return (rng.standard_exponential(size=size) - rng.standard_exponential(size=size)) * self.s
+        r = rng or nr
+        return (r.standard_exponential(size=size) - r.standard_exponential(size=size)) * self.s
 
 
 class PoissonProposal(Proposal):
     def __call__(self, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
-        return rng.poisson(lam=self.s, size=np.size(self.s)) - self.s
+        return (rng or nr).poisson(lam=self.s, size=np.size(self.s)) - self.s
 
 
 class MultivariateNormalProposal(Proposal):
@@ -100,13 +102,12 @@ class MultivariateNormalProposal(Proposal):
         self.chol = scipy.linalg.cholesky(s, lower=True)
 
     def __call__(self, num_draws=None, rng: Optional[np.random.Generator] = None):
-        if rng is None:
-            rng = nr
+        rng_ = rng or nr
         if num_draws is not None:
-            b = rng.normal(size=(self.n, num_draws))
+            b = rng_.normal(size=(self.n, num_draws))
             return np.dot(self.chol, b).T
         else:
-            b = rng.normal(size=self.n)
+            b = rng_.normal(size=self.n)
             return np.dot(self.chol, b)
 
 
@@ -116,15 +117,12 @@ class Metropolis(ArrayStepShared):
     name = "metropolis"
 
     default_blocked = False
-    generates_stats = True
-    stats_dtypes = [
-        {
-            "accept": np.float64,
-            "accepted": bool,
-            "tune": bool,
-            "scaling": np.float64,
-        }
-    ]
+    stats_dtypes_shapes = {
+        "accept": (np.float64, []),
+        "accepted": (np.float64, []),
+        "tune": (bool, []),
+        "scaling": (np.float64, []),
+    }
 
     def __init__(
         self,
@@ -158,20 +156,20 @@ class Metropolis(ArrayStepShared):
         model: PyMC Model
             Optional model for sampling step. Defaults to None (taken from context).
         mode: string or `Mode` instance.
-            compilation mode passed to Aesara functions
+            compilation mode passed to PyTensor functions
         """
 
         model = pm.modelcontext(model)
-        initial_values = model.compute_initial_point()
+        initial_values = model.initial_point()
 
         if vars is None:
             vars = model.value_vars
         else:
-            vars = [model.rvs_to_values.get(var, var) for var in vars]
-        vars = pm.inputvars(vars)
+            vars = get_value_vars_from_user_vars(vars, model)
 
+        initial_values_shape = [initial_values[v.name].shape for v in vars]
         if S is None:
-            S = np.ones(sum(initial_values[v.name].size for v in vars))
+            S = np.ones(int(sum(np.prod(ivs) for ivs in initial_values_shape)))
 
         if proposal_dist is not None:
             self.proposal_dist = proposal_dist(S)
@@ -186,7 +184,6 @@ class Metropolis(ArrayStepShared):
         self.tune = tune
         self.tune_interval = tune_interval
         self.steps_until_tune = tune_interval
-        self.accepted = 0
 
         # Determine type of variables
         self.discrete = np.concatenate(
@@ -195,65 +192,101 @@ class Metropolis(ArrayStepShared):
         self.any_discrete = self.discrete.any()
         self.all_discrete = self.discrete.all()
 
-        # remember initial settings before tuning so they can be reset
-        self._untuned_settings = dict(
-            scaling=self.scaling, steps_until_tune=tune_interval, accepted=self.accepted
+        # Metropolis will try to handle one batched dimension at a time This, however,
+        # is not safe for discrete multivariate distributions (looking at you Multinomial),
+        # due to high dependency among the support dimensions. For continuous multivariate
+        # distributions we assume they are being transformed in a way that makes each
+        # dimension semi-independent.
+        is_scalar = len(initial_values_shape) == 1 and initial_values_shape[0] == ()
+        self.elemwise_update = not (
+            is_scalar
+            or (
+                self.any_discrete
+                and max(getattr(model.values_to_rvs[var].owner.op, "ndim_supp", 1) for var in vars)
+                > 0
+            )
         )
+        if self.elemwise_update:
+            dims = int(sum(np.prod(ivs) for ivs in initial_values_shape))
+        else:
+            dims = 1
+        self.enum_dims = np.arange(dims, dtype=int)
+        self.accept_rate_iter = np.zeros(dims, dtype=float)
+        self.accepted_iter = np.zeros(dims, dtype=bool)
+        self.accepted_sum = np.zeros(dims, dtype=int)
 
+        # remember initial settings before tuning so they can be reset
+        self._untuned_settings = dict(scaling=self.scaling, steps_until_tune=tune_interval)
+
+        # TODO: This is not being used when compiling the logp function!
         self.mode = mode
 
         shared = pm.make_shared_replacements(initial_values, vars, model)
-        self.delta_logp = delta_logp(initial_values, model.logpt(), vars, shared)
+        self.delta_logp = delta_logp(initial_values, model.logp(), vars, shared)
         super().__init__(vars, shared)
 
     def reset_tuning(self):
         """Resets the tuned sampler parameters to their initial values."""
         for attr, initial_value in self._untuned_settings.items():
             setattr(self, attr, initial_value)
+        self.accepted_sum[:] = 0
         return
 
-    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, List[Dict[str, Any]]]:
-
+    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, StatsType]:
         point_map_info = q0.point_map_info
-        q0 = q0.data
+        q0d = q0.data
 
         if not self.steps_until_tune and self.tune:
             # Tune scaling parameter
-            self.scaling = tune(self.scaling, self.accepted / float(self.tune_interval))
+            self.scaling = tune(self.scaling, self.accepted_sum / float(self.tune_interval))
             # Reset counter
             self.steps_until_tune = self.tune_interval
-            self.accepted = 0
+            self.accepted_sum[:] = 0
 
         delta = self.proposal_dist() * self.scaling
 
         if self.any_discrete:
             if self.all_discrete:
                 delta = np.round(delta, 0).astype("int64")
-                q0 = q0.astype("int64")
-                q = (q0 + delta).astype("int64")
+                q0d = q0d.astype("int64")
+                q = (q0d + delta).astype("int64")
             else:
                 delta[self.discrete] = np.round(delta[self.discrete], 0)
-                q = q0 + delta
+                q = q0d + delta
         else:
-            q = floatX(q0 + delta)
+            q = floatX(q0d + delta)
 
-        accept = self.delta_logp(q, q0)
-        q_new, accepted = metrop_select(accept, q, q0)
-
-        self.accepted += accepted
+        if self.elemwise_update:
+            q0d = q0d.copy()
+            q_temp = q0d.copy()
+            # Shuffle order of updates (probably we don't need to do this in every step)
+            np.random.shuffle(self.enum_dims)
+            for i in self.enum_dims:
+                q_temp[i] = q[i]
+                accept_rate_i = self.delta_logp(q_temp, q0d)
+                q_temp_, accepted_i = metrop_select(accept_rate_i, q_temp, q0d)
+                q_temp[i] = q0d[i] = q_temp_[i]
+                self.accept_rate_iter[i] = accept_rate_i
+                self.accepted_iter[i] = accepted_i
+                self.accepted_sum[i] += accepted_i
+            q = q_temp
+        else:
+            accept_rate = self.delta_logp(q, q0d)
+            q, accepted = metrop_select(accept_rate, q, q0d)
+            self.accept_rate_iter = accept_rate
+            self.accepted_iter = accepted
+            self.accepted_sum += accepted
 
         self.steps_until_tune -= 1
 
         stats = {
             "tune": self.tune,
-            "scaling": self.scaling,
-            "accept": np.exp(accept),
-            "accepted": accepted,
+            "scaling": np.mean(self.scaling),
+            "accept": np.mean(np.exp(self.accept_rate_iter)),
+            "accepted": np.mean(self.accepted_iter),
         }
 
-        q_new = RaveledVars(q_new, point_map_info)
-
-        return q_new, [stats]
+        return RaveledVars(q, point_map_info), [stats]
 
     @staticmethod
     def competence(var, has_grad):
@@ -275,26 +308,38 @@ def tune(scale, acc_rate):
     >0.95         x 10
 
     """
-    if acc_rate < 0.001:
+    return scale * np.where(
+        acc_rate < 0.001,
         # reduce by 90 percent
-        return scale * 0.1
-    elif acc_rate < 0.05:
-        # reduce by 50 percent
-        return scale * 0.5
-    elif acc_rate < 0.2:
-        # reduce by ten percent
-        return scale * 0.9
-    elif acc_rate > 0.95:
-        # increase by factor of ten
-        return scale * 10.0
-    elif acc_rate > 0.75:
-        # increase by double
-        return scale * 2.0
-    elif acc_rate > 0.5:
-        # increase by ten percent
-        return scale * 1.1
-
-    return scale
+        0.1,
+        np.where(
+            acc_rate < 0.05,
+            # reduce by 50 percent
+            0.5,
+            np.where(
+                acc_rate < 0.2,
+                # reduce by ten percent
+                0.9,
+                np.where(
+                    acc_rate > 0.95,
+                    # increase by factor of ten
+                    10.0,
+                    np.where(
+                        acc_rate > 0.75,
+                        # increase by double
+                        2.0,
+                        np.where(
+                            acc_rate > 0.5,
+                            # increase by ten percent
+                            1.1,
+                            # Do not change
+                            1.0,
+                        ),
+                    ),
+                ),
+            ),
+        ),
+    )
 
 
 class BinaryMetropolis(ArrayStep):
@@ -317,17 +362,13 @@ class BinaryMetropolis(ArrayStep):
 
     name = "binary_metropolis"
 
-    generates_stats = True
-    stats_dtypes = [
-        {
-            "accept": np.float64,
-            "tune": bool,
-            "p_jump": np.float64,
-        }
-    ]
+    stats_dtypes_shapes = {
+        "accept": (np.float64, []),
+        "tune": (bool, []),
+        "p_jump": (np.float64, []),
+    }
 
     def __init__(self, vars, scaling=1.0, tune=True, tune_interval=100, model=None):
-
         model = pm.modelcontext(model)
 
         self.scaling = scaling
@@ -336,18 +377,18 @@ class BinaryMetropolis(ArrayStep):
         self.steps_until_tune = tune_interval
         self.accepted = 0
 
-        vars = [model.rvs_to_values.get(var, var) for var in vars]
+        vars = get_value_vars_from_user_vars(vars, model)
 
         if not all([v.dtype in pm.discrete_types for v in vars]):
             raise ValueError("All variables must be Bernoulli for BinaryMetropolis")
 
         super().__init__(vars, [model.compile_logp()])
 
-    def astep(self, q0: RaveledVars, logp) -> Tuple[RaveledVars, List[Dict[str, Any]]]:
-
-        logp_q0 = logp(q0)
-        point_map_info = q0.point_map_info
-        q0 = q0.data
+    def astep(self, apoint: RaveledVars, *args) -> Tuple[RaveledVars, StatsType]:
+        logp = args[0]
+        logp_q0 = logp(apoint)
+        point_map_info = apoint.point_map_info
+        q0 = apoint.data
 
         # Convert adaptive_scale_factor to a jump probability
         p_jump = 1.0 - 0.5**self.scaling
@@ -369,9 +410,7 @@ class BinaryMetropolis(ArrayStep):
             "p_jump": p_jump,
         }
 
-        q_new = RaveledVars(q_new, point_map_info)
-
-        return q_new, [stats]
+        return RaveledVars(q_new, point_map_info), [stats]
 
     @staticmethod
     def competence(var):
@@ -419,15 +458,22 @@ class BinaryGibbsMetropolis(ArrayStep):
 
     name = "binary_gibbs_metropolis"
 
-    def __init__(self, vars, order="random", transit_p=0.8, model=None):
+    stats_dtypes_shapes = {
+        "tune": (bool, []),
+    }
 
+    def __init__(self, vars, order="random", transit_p=0.8, model=None):
         model = pm.modelcontext(model)
 
+        # Doesn't actually tune, but it's required to emit a sampler stat
+        # that indicates whether a draw was done in a tuning phase.
+        self.tune = True
         # transition probabilities
         self.transit_p = transit_p
 
-        initial_point = model.compute_initial_point()
-        vars = [model.rvs_to_values.get(var, var) for var in vars]
+        vars = get_value_vars_from_user_vars(vars, model)
+
+        initial_point = model.initial_point()
         self.dim = sum(initial_point[v.name].size for v in vars)
 
         if order == "random":
@@ -444,13 +490,18 @@ class BinaryGibbsMetropolis(ArrayStep):
 
         super().__init__(vars, [model.compile_logp()])
 
-    def astep(self, q0: RaveledVars, logp: Callable[[RaveledVars], np.ndarray]) -> RaveledVars:
+    def reset_tuning(self):
+        # There are no tuning parameters in this step method.
+        self.tune = False
+        return
 
+    def astep(self, apoint: RaveledVars, *args) -> Tuple[RaveledVars, StatsType]:
+        logp: Callable[[RaveledVars], np.ndarray] = args[0]
         order = self.order
         if self.shuffle_dims:
             nr.shuffle(order)
 
-        q = RaveledVars(np.copy(q0.data), q0.point_map_info)
+        q = RaveledVars(np.copy(apoint.data), apoint.point_map_info)
 
         logp_curr = logp(q)
 
@@ -464,7 +515,10 @@ class BinaryGibbsMetropolis(ArrayStep):
                 if accepted:
                     logp_curr = logp_prop
 
-        return q
+        stats = {
+            "tune": self.tune,
+        }
+        return q, [stats]
 
     @staticmethod
     def competence(var):
@@ -504,14 +558,16 @@ class CategoricalGibbsMetropolis(ArrayStep):
 
     name = "categorical_gibbs_metropolis"
 
-    def __init__(self, vars, proposal="uniform", order="random", model=None):
+    stats_dtypes_shapes = {
+        "tune": (bool, []),
+    }
 
+    def __init__(self, vars, proposal="uniform", order="random", model=None):
         model = pm.modelcontext(model)
 
-        vars = [model.rvs_to_values.get(var, var) for var in vars]
-        vars = pm.inputvars(vars)
+        vars = get_value_vars_from_user_vars(vars, model)
 
-        initial_point = model.compute_initial_point()
+        initial_point = model.initial_point()
 
         dimcats = []
         # The above variable is a list of pairs (aggregate dimension, number
@@ -519,7 +575,6 @@ class CategoricalGibbsMetropolis(ArrayStep):
         # variable with M categories and y being a 3-D variable with N
         # categories, we will have dimcats = [(0, M), (1, M), (2, N), (3, N), (4, N)].
         for v in vars:
-
             v_init_val = initial_point[v.name]
 
             rv_var = model.values_to_rvs[v]
@@ -527,7 +582,7 @@ class CategoricalGibbsMetropolis(ArrayStep):
 
             if isinstance(distr, CategoricalRV):
                 k_graph = rv_var.owner.inputs[3].shape[-1]
-                (k_graph,), _ = rvs_to_value_vars((k_graph,), apply_transforms=True)
+                (k_graph,) = model.replace_rvs_by_values((k_graph,))
                 k = model.compile_fn(k_graph, inputs=model.value_vars, on_unused_input="ignore")(
                     initial_point
                 )
@@ -557,12 +612,21 @@ class CategoricalGibbsMetropolis(ArrayStep):
         else:
             raise ValueError("Argument 'proposal' should either be 'uniform' or 'proportional'")
 
+        # Doesn't actually tune, but it's required to emit a sampler stat
+        # that indicates whether a draw was done in a tuning phase.
+        self.tune = True
+
         super().__init__(vars, [model.compile_logp()])
 
-    def astep_unif(self, q0: RaveledVars, logp) -> RaveledVars:
+    def reset_tuning(self):
+        # There are no tuning parameters in this step method.
+        self.tune = False
+        return
 
-        point_map_info = q0.point_map_info
-        q0 = q0.data
+    def astep_unif(self, apoint: RaveledVars, *args) -> Tuple[RaveledVars, StatsType]:
+        logp = args[0]
+        point_map_info = apoint.point_map_info
+        q0 = apoint.data
 
         dimcats = self.dimcats
         if self.shuffle_dims:
@@ -578,12 +642,15 @@ class CategoricalGibbsMetropolis(ArrayStep):
             if accepted:
                 logp_curr = logp_prop
 
-        return q
+        stats = {
+            "tune": self.tune,
+        }
+        return q, [stats]
 
-    def astep_prop(self, q0: RaveledVars, logp) -> RaveledVars:
-
-        point_map_info = q0.point_map_info
-        q0 = q0.data
+    def astep_prop(self, apoint: RaveledVars, *args) -> Tuple[RaveledVars, StatsType]:
+        logp = args[0]
+        point_map_info = apoint.point_map_info
+        q0 = apoint.data
 
         dimcats = self.dimcats
         if self.shuffle_dims:
@@ -595,9 +662,9 @@ class CategoricalGibbsMetropolis(ArrayStep):
         for dim, k in dimcats:
             logp_curr = self.metropolis_proportional(q, logp, logp_curr, dim, k)
 
-        return q
+        return q, []
 
-    def astep(self, q0, logp):
+    def astep(self, apoint: RaveledVars, *args) -> Tuple[RaveledVars, StatsType]:
         raise NotImplementedError()
 
     def metropolis_proportional(self, q, logp, logp_curr, dim, k):
@@ -663,17 +730,17 @@ class DEMetropolis(PopulationArrayStepShared):
         Some measure of variance to parameterize proposal distribution
     proposal_dist: function
         Function that returns zero-mean deviates when parameterized with
-        S (and n). Defaults to Uniform(-S,+S).
+        S (and n). Defaults to NormalProposal(S).
     scaling: scalar or array
         Initial scale factor for epsilon. Defaults to 0.001
     tune: str
-        Which hyperparameter to tune. Defaults to None, but can also be 'scaling' or 'lambda'.
+        Which hyperparameter to tune. Defaults to 'scaling', but can also be 'lambda' or None.
     tune_interval: int
         The frequency of tuning. Defaults to 100 iterations.
     model: PyMC Model
         Optional model for sampling step. Defaults to None (taken from context).
     mode:  string or `Mode` instance.
-        compilation mode passed to Aesara functions
+        compilation mode passed to PyTensor functions
 
     References
     ----------
@@ -687,16 +754,13 @@ class DEMetropolis(PopulationArrayStepShared):
     name = "DEMetropolis"
 
     default_blocked = True
-    generates_stats = True
-    stats_dtypes = [
-        {
-            "accept": np.float64,
-            "accepted": bool,
-            "tune": bool,
-            "scaling": np.float64,
-            "lambda": np.float64,
-        }
-    ]
+    stats_dtypes_shapes = {
+        "accept": (np.float64, []),
+        "accepted": (bool, []),
+        "tune": (bool, []),
+        "scaling": (np.float64, []),
+        "lambda": (np.float64, []),
+    }
 
     def __init__(
         self,
@@ -705,22 +769,20 @@ class DEMetropolis(PopulationArrayStepShared):
         proposal_dist=None,
         lamb=None,
         scaling=0.001,
-        tune=None,
+        tune: Optional[str] = "scaling",
         tune_interval=100,
         model=None,
         mode=None,
         **kwargs
     ):
-
         model = pm.modelcontext(model)
-        initial_values = model.compute_initial_point()
+        initial_values = model.initial_point()
         initial_values_size = sum(initial_values[n.name].size for n in model.value_vars)
 
         if vars is None:
-            vars = model.cont_vars
+            vars = model.continuous_value_vars
         else:
-            vars = [model.rvs_to_values.get(var, var) for var in vars]
-        vars = pm.inputvars(vars)
+            vars = get_value_vars_from_user_vars(vars, model)
 
         if S is None:
             S = np.ones(initial_values_size)
@@ -728,7 +790,7 @@ class DEMetropolis(PopulationArrayStepShared):
         if proposal_dist is not None:
             self.proposal_dist = proposal_dist(S)
         else:
-            self.proposal_dist = UniformProposal(S)
+            self.proposal_dist = NormalProposal(S)
 
         self.scaling = np.atleast_1d(scaling).astype("d")
         if lamb is None:
@@ -745,13 +807,12 @@ class DEMetropolis(PopulationArrayStepShared):
         self.mode = mode
 
         shared = pm.make_shared_replacements(initial_values, vars, model)
-        self.delta_logp = delta_logp(initial_values, model.logpt(), vars, shared)
+        self.delta_logp = delta_logp(initial_values, model.logp(), vars, shared)
         super().__init__(vars, shared)
 
-    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, List[Dict[str, Any]]]:
-
+    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, StatsType]:
         point_map_info = q0.point_map_info
-        q0 = q0.data
+        q0d = q0.data
 
         if not self.steps_until_tune and self.tune:
             if self.tune == "scaling":
@@ -770,10 +831,10 @@ class DEMetropolis(PopulationArrayStepShared):
         r1 = DictToArrayBijection.map(self.population[ir1])
         r2 = DictToArrayBijection.map(self.population[ir2])
         # propose a jump
-        q = floatX(q0 + self.lamb * (r1.data - r2.data) + epsilon)
+        q = floatX(q0d + self.lamb * (r1.data - r2.data) + epsilon)
 
-        accept = self.delta_logp(q, q0)
-        q_new, accepted = metrop_select(accept, q, q0)
+        accept = self.delta_logp(q, q0d)
+        q_new, accepted = metrop_select(accept, q, q0d)
         self.accepted += accepted
 
         self.steps_until_tune -= 1
@@ -786,9 +847,7 @@ class DEMetropolis(PopulationArrayStepShared):
             "accepted": accepted,
         }
 
-        q_new = RaveledVars(q_new, point_map_info)
-
-        return q_new, [stats]
+        return RaveledVars(q_new, point_map_info), [stats]
 
     @staticmethod
     def competence(var, has_grad):
@@ -811,11 +870,11 @@ class DEMetropolisZ(ArrayStepShared):
         Some measure of variance to parameterize proposal distribution
     proposal_dist: function
         Function that returns zero-mean deviates when parameterized with
-        S (and n). Defaults to Uniform(-S,+S).
+        S (and n). Defaults to NormalProposal(S).
     scaling: scalar or array
         Initial scale factor for epsilon. Defaults to 0.001
     tune: str
-        Which hyperparameter to tune. Defaults to 'lambda', but can also be 'scaling' or None.
+        Which hyperparameter to tune. Defaults to 'scaling', but can also be 'lambda' or None.
     tune_interval: int
         The frequency of tuning. Defaults to 100 iterations.
     tune_drop_fraction: float
@@ -825,11 +884,11 @@ class DEMetropolisZ(ArrayStepShared):
     model: PyMC Model
         Optional model for sampling step. Defaults to None (taken from context).
     mode:  string or `Mode` instance.
-        compilation mode passed to Aesara functions
+        compilation mode passed to PyTensor functions
 
     References
     ----------
-    .. [Braak2006] Cajo C.F. ter Braak (2006).
+    .. [Braak2008] Cajo C.F. ter Braak (2008).
         Differential Evolution Markov Chain with snooker updater and fewer chains.
         Statistics and Computing
         `link <https://doi.org/10.1007/s11222-008-9104-9>`__
@@ -838,16 +897,13 @@ class DEMetropolisZ(ArrayStepShared):
     name = "DEMetropolisZ"
 
     default_blocked = True
-    generates_stats = True
-    stats_dtypes = [
-        {
-            "accept": np.float64,
-            "accepted": bool,
-            "tune": bool,
-            "scaling": np.float64,
-            "lambda": np.float64,
-        }
-    ]
+    stats_dtypes_shapes = {
+        "accept": (np.float64, []),
+        "accepted": (bool, []),
+        "tune": (bool, []),
+        "scaling": (np.float64, []),
+        "lambda": (np.float64, []),
+    }
 
     def __init__(
         self,
@@ -856,7 +912,7 @@ class DEMetropolisZ(ArrayStepShared):
         proposal_dist=None,
         lamb=None,
         scaling=0.001,
-        tune="lambda",
+        tune: Optional[str] = "scaling",
         tune_interval=100,
         tune_drop_fraction: float = 0.9,
         model=None,
@@ -864,14 +920,13 @@ class DEMetropolisZ(ArrayStepShared):
         **kwargs
     ):
         model = pm.modelcontext(model)
-        initial_values = model.compute_initial_point()
+        initial_values = model.initial_point()
         initial_values_size = sum(initial_values[n.name].size for n in model.value_vars)
 
         if vars is None:
-            vars = model.cont_vars
+            vars = model.continuous_value_vars
         else:
-            vars = [model.rvs_to_values.get(var, var) for var in vars]
-        vars = pm.inputvars(vars)
+            vars = get_value_vars_from_user_vars(vars, model)
 
         if S is None:
             S = np.ones(initial_values_size)
@@ -879,7 +934,7 @@ class DEMetropolisZ(ArrayStepShared):
         if proposal_dist is not None:
             self.proposal_dist = proposal_dist(S)
         else:
-            self.proposal_dist = UniformProposal(S)
+            self.proposal_dist = NormalProposal(S)
 
         self.scaling = np.atleast_1d(scaling).astype("d")
         if lamb is None:
@@ -896,7 +951,7 @@ class DEMetropolisZ(ArrayStepShared):
         self.accepted = 0
 
         # cache local history for the Z-proposals
-        self._history = []
+        self._history: List[np.ndarray] = []
         # remember initial settings before tuning so they can be reset
         self._untuned_settings = dict(
             scaling=self.scaling,
@@ -908,7 +963,7 @@ class DEMetropolisZ(ArrayStepShared):
         self.mode = mode
 
         shared = pm.make_shared_replacements(initial_values, vars, model)
-        self.delta_logp = delta_logp(initial_values, model.logpt(), vars, shared)
+        self.delta_logp = delta_logp(initial_values, model.logp(), vars, shared)
         super().__init__(vars, shared)
 
     def reset_tuning(self):
@@ -919,10 +974,9 @@ class DEMetropolisZ(ArrayStepShared):
             setattr(self, attr, initial_value)
         return
 
-    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, List[Dict[str, Any]]]:
-
+    def astep(self, q0: RaveledVars) -> Tuple[RaveledVars, StatsType]:
         point_map_info = q0.point_map_info
-        q0 = q0.data
+        q0d = q0.data
 
         # same tuning scheme as DEMetropolis
         if not self.steps_until_tune and self.tune:
@@ -949,13 +1003,13 @@ class DEMetropolisZ(ArrayStepShared):
             z1 = self._history[iz1]
             z2 = self._history[iz2]
             # propose a jump
-            q = floatX(q0 + self.lamb * (z1 - z2) + epsilon)
+            q = floatX(q0d + self.lamb * (z1 - z2) + epsilon)
         else:
             # propose just with noise in the first 2 iterations
-            q = floatX(q0 + epsilon)
+            q = floatX(q0d + epsilon)
 
-        accept = self.delta_logp(q, q0)
-        q_new, accepted = metrop_select(accept, q, q0)
+        accept = self.delta_logp(q, q0d)
+        q_new, accepted = metrop_select(accept, q, q0d)
         self.accepted += accepted
         self._history.append(q_new)
 
@@ -969,9 +1023,7 @@ class DEMetropolisZ(ArrayStepShared):
             "accepted": accepted,
         }
 
-        q_new = RaveledVars(q_new, point_map_info)
-
-        return q_new, [stats]
+        return RaveledVars(q_new, point_map_info), [stats]
 
     def stop_tuning(self):
         """At the end of the tuning phase, this method removes the first x% of the history
@@ -996,13 +1048,22 @@ def sample_except(limit, excluded):
     return candidate
 
 
-def delta_logp(point, logp, vars, shared):
-    [logp0], inarray0 = pm.join_nonshared_inputs(point, [logp], vars, shared)
+def delta_logp(
+    point: Dict[str, np.ndarray],
+    logp: pt.TensorVariable,
+    vars: List[pt.TensorVariable],
+    shared: Dict[pt.TensorVariable, pt.sharedvar.TensorSharedVariable],
+) -> pytensor.compile.Function:
+    [logp0], inarray0 = join_nonshared_inputs(
+        point=point, outputs=[logp], inputs=vars, shared_inputs=shared
+    )
 
     tensor_type = inarray0.type
     inarray1 = tensor_type("inarray1")
 
-    logp1 = pm.CallableTensor(logp0)(inarray1)
+    logp1 = CallableTensor(logp0)(inarray1)
+    # Replace any potential duplicated RNG nodes
+    (logp1,) = replace_rng_nodes((logp1,))
 
     f = compile_pymc([inarray1, inarray0], logp1 - logp0)
     f.trust_input = True
