@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -14,13 +14,17 @@
 
 # Modified from original implementation by Dominik Wabersich (2013)
 
+from typing import Tuple
+
 import numpy as np
 import numpy.random as nr
 
-from pymc.aesaraf import inputvars
-from pymc.blocking import RaveledVars
+from pymc.blocking import RaveledVars, StatsType
 from pymc.model import modelcontext
-from pymc.step_methods.arraystep import ArrayStep, Competence
+from pymc.pytensorf import compile_pymc, join_nonshared_inputs, make_shared_replacements
+from pymc.step_methods.arraystep import ArrayStepShared
+from pymc.step_methods.compound import Competence
+from pymc.util import get_value_vars_from_user_vars
 from pymc.vartypes import continuous_types
 
 __all__ = ["Slice"]
@@ -28,9 +32,9 @@ __all__ = ["Slice"]
 LOOP_ERR_MSG = "max slicer iters %d exceeded"
 
 
-class Slice(ArrayStep):
+class Slice(ArrayStepShared):
     """
-    Univariate slice sampler step method
+    Univariate slice sampler step method.
 
     Parameters
     ----------
@@ -47,53 +51,76 @@ class Slice(ArrayStep):
 
     name = "slice"
     default_blocked = False
+    stats_dtypes_shapes = {
+        "tune": (bool, []),
+        "nstep_out": (int, []),
+        "nstep_in": (int, []),
+    }
 
     def __init__(self, vars=None, w=1.0, tune=True, model=None, iter_limit=np.inf, **kwargs):
-        self.model = modelcontext(model)
-        self.w = w
+        model = modelcontext(model)
+        self.w = np.asarray(w).copy()
         self.tune = tune
         self.n_tunes = 0.0
         self.iter_limit = iter_limit
 
         if vars is None:
-            vars = self.model.cont_vars
+            vars = model.continuous_value_vars
         else:
-            vars = [self.model.rvs_to_values.get(var, var) for var in vars]
-        vars = inputvars(vars)
+            vars = get_value_vars_from_user_vars(vars, model)
 
-        super().__init__(vars, [self.model.compile_logp()], **kwargs)
+        point = model.initial_point()
+        shared = make_shared_replacements(point, vars, model)
+        [logp], raveled_inp = join_nonshared_inputs(
+            point=point, outputs=[model.logp()], inputs=vars, shared_inputs=shared
+        )
+        self.logp = compile_pymc([raveled_inp], logp)
+        self.logp.trust_input = True
 
-    def astep(self, q0, logp):
-        q0_val = q0.data
-        self.w = np.resize(self.w, len(q0_val))  # this is a repmat
-        q = np.copy(q0_val)  # TODO: find out if we need this
+        super().__init__(vars, shared)
+
+    def astep(self, apoint: RaveledVars) -> Tuple[RaveledVars, StatsType]:
+        # The arguments are determined by the list passed via `super().__init__(..., fs, ...)`
+        q0_val = apoint.data
+
+        if q0_val.shape != self.w.shape:
+            self.w = np.resize(self.w, len(q0_val))  # this is a repmat
+
+        nstep_out = nstep_in = 0
+
+        q = np.copy(q0_val)
         ql = np.copy(q0_val)  # l for left boundary
-        qr = np.copy(q0_val)  # r for right boudary
-        for i in range(len(q0_val)):
+        qr = np.copy(q0_val)  # r for right boundary
+
+        logp = self.logp
+        for i, wi in enumerate(self.w):
             # uniformly sample from 0 to p(q), but in log space
-            q_ra = RaveledVars(q, q0.point_map_info)
-            y = logp(q_ra) - nr.standard_exponential()
-            ql[i] = q[i] - nr.uniform(0, self.w[i])
-            qr[i] = q[i] + self.w[i]
+            y = logp(q) - nr.standard_exponential()
+
+            # Create initial interval
+            ql[i] = q[i] - nr.uniform() * wi  # q[i] + r * w
+            qr[i] = ql[i] + wi  # Equivalent to q[i] + (1-r) * w
+
             # Stepping out procedure
             cnt = 0
-            while y <= logp(
-                RaveledVars(ql, q0.point_map_info)
-            ):  # changed lt to leq  for locally uniform posteriors
-                ql[i] -= self.w[i]
+            while y <= logp(ql):  # changed lt to leq  for locally uniform posteriors
+                ql[i] -= wi
                 cnt += 1
                 if cnt > self.iter_limit:
                     raise RuntimeError(LOOP_ERR_MSG % self.iter_limit)
+            nstep_out += cnt
+
             cnt = 0
-            while y <= logp(RaveledVars(qr, q0.point_map_info)):
-                qr[i] += self.w[i]
+            while y <= logp(qr):
+                qr[i] += wi
                 cnt += 1
                 if cnt > self.iter_limit:
                     raise RuntimeError(LOOP_ERR_MSG % self.iter_limit)
+            nstep_out += cnt
 
             cnt = 0
             q[i] = nr.uniform(ql[i], qr[i])
-            while logp(q_ra) < y:  # Changed leq to lt, to accomodate for locally flat posteriors
+            while y > logp(q):  # Changed leq to lt, to accommodate for locally flat posteriors
                 # Sample uniformly from slice
                 if q[i] > q0_val[i]:
                     qr[i] = q[i]
@@ -103,20 +130,28 @@ class Slice(ArrayStep):
                 cnt += 1
                 if cnt > self.iter_limit:
                     raise RuntimeError(LOOP_ERR_MSG % self.iter_limit)
+            nstep_in += cnt
 
-            if (
-                self.tune
-            ):  # I was under impression from MacKays lectures that slice width can be tuned without
+            if self.tune:
+                # I was under impression from MacKays lectures that slice width can be tuned without
                 # breaking markovianness. Can we do it regardless of self.tune?(@madanh)
-                self.w[i] = self.w[i] * (self.n_tunes / (self.n_tunes + 1)) + (qr[i] - ql[i]) / (
+                self.w[i] = wi * (self.n_tunes / (self.n_tunes + 1)) + (qr[i] - ql[i]) / (
                     self.n_tunes + 1
-                )  # same as before
-                # unobvious and important: return qr and ql to the same point
-                qr[i] = q[i]
-                ql[i] = q[i]
+                )
+
+            # Set qr and ql to the accepted points (they matter for subsequent iterations)
+            qr[i] = ql[i] = q[i]
+
         if self.tune:
             self.n_tunes += 1
-        return q
+
+        stats = {
+            "tune": self.tune,
+            "nstep_out": nstep_out,
+            "nstep_in": nstep_in,
+        }
+
+        return RaveledVars(q, apoint.point_map_info), [stats]
 
     @staticmethod
     def competence(var, has_grad):
