@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -16,7 +16,6 @@
 from pymc.variational import (
     approximations,
     callbacks,
-    flows,
     inference,
     operators,
     opvi,
@@ -27,13 +26,11 @@ from pymc.variational.approximations import (
     Empirical,
     FullRank,
     MeanField,
-    NormalizingFlow,
     sample_approx,
 )
 from pymc.variational.inference import (
     ADVI,
     ASVGD,
-    NFVI,
     SVGD,
     FullRankADVI,
     ImplicitGradient,
