@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -45,24 +45,45 @@ References
     https://arxiv.org/abs/1610.09033 (2016)
 """
 
+from __future__ import annotations
+
 import collections
 import itertools
 import warnings
 
-import aesara
-import aesara.tensor as at
-import numpy as np
+from typing import Any
 
-from aesara.graph.basic import Variable
+import numpy as np
+import pytensor
+import pytensor.tensor as pt
+import xarray
+
+from pytensor.graph.basic import Variable
+from pytensor.graph.replace import graph_replace
+from pytensor.tensor.shape import unbroadcast
 
 import pymc as pm
 
-from pymc.aesaraf import at_rng, compile_pymc, identity, rvs_to_value_vars
-from pymc.backends import NDArray
+from pymc.backends.base import MultiTrace
+from pymc.backends.ndarray import NDArray
 from pymc.blocking import DictToArrayBijection
 from pymc.initial_point import make_initial_point_fn
 from pymc.model import modelcontext
-from pymc.util import WithMemoization, locally_cachedmethod
+from pymc.pytensorf import (
+    SeedSequenceSeed,
+    compile_pymc,
+    find_rng_nodes,
+    identity,
+    reseed_rngs,
+)
+from pymc.util import (
+    RandomState,
+    WithMemoization,
+    _get_seeds_per_chain,
+    locally_cachedmethod,
+    makeiter,
+)
+from pymc.variational.minibatch_rv import MinibatchRandomVariable, get_scaling
 from pymc.variational.updates import adagrad_window
 from pymc.vartypes import discrete_types
 
@@ -93,12 +114,16 @@ class GroupError(VariationalInferenceError, TypeError):
     """Error related to VI groups"""
 
 
-class BatchedGroupError(GroupError):
-    """Error with batched variables"""
+def _known_scan_ignored_inputs(terms):
+    # TODO: remove when scan issue with grads is fixed
+    from pymc.data import MinibatchIndexRV
+    from pymc.distributions.simulator import SimulatorRV
 
-
-class LocalGroupError(BatchedGroupError, AEVBInferenceError):
-    """Error raised in case of bad local_rv usage"""
+    return [
+        n.owner.inputs[0]
+        for n in pytensor.graph.ancestors(terms)
+        if n.owner is not None and isinstance(n.owner.op, (MinibatchIndexRV, SimulatorRV))
+    ]
 
 
 def append_name(name):
@@ -123,21 +148,21 @@ def node_property(f):
 
         def wrapper(fn):
             ff = append_name(f)(fn)
-            f_ = aesara.config.change_flags(compute_test_value="off")(ff)
+            f_ = pytensor.config.change_flags(compute_test_value="off")(ff)
             return property(locally_cachedmethod(f_))
 
         return wrapper
     else:
-        f_ = aesara.config.change_flags(compute_test_value="off")(f)
+        f_ = pytensor.config.change_flags(compute_test_value="off")(f)
         return property(locally_cachedmethod(f_))
 
 
-@aesara.config.change_flags(compute_test_value="ignore")
+@pytensor.config.change_flags(compute_test_value="ignore")
 def try_to_set_test_value(node_in, node_out, s):
     _s = s
     if s is None:
         s = 1
-    s = aesara.compile.view_op(at.as_tensor(s))
+    s = pytensor.compile.view_op(pt.as_tensor(s))
     if not isinstance(node_in, (list, tuple)):
         node_in = [node_in]
     if not isinstance(node_out, (list, tuple)):
@@ -154,7 +179,7 @@ def try_to_set_test_value(node_in, node_out, s):
                 o.tag.test_value = tv
 
 
-class ObjectiveUpdates(aesara.OrderedUpdates):
+class ObjectiveUpdates(pytensor.OrderedUpdates):
     """OrderedUpdates extension for storing loss"""
 
     loss = None
@@ -175,7 +200,7 @@ class ObjectiveFunction:
         OPVI TestFunction
     """
 
-    def __init__(self, op, tf):
+    def __init__(self, op: Operator, tf: TestFunction):
         self.op = op
         self.tf = tf
 
@@ -295,7 +320,7 @@ class ObjectiveFunction:
         if self.op.returns_loss:
             updates.loss = obj_target
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def step_function(
         self,
         obj_n_mc=None,
@@ -339,13 +364,13 @@ class ObjectiveFunction:
         score: `bool`
             calculate loss on each step? Defaults to False for speed
         fn_kwargs: `dict`
-            Add kwargs to aesara.function (e.g. `{'profile': True}`)
+            Add kwargs to pytensor.function (e.g. `{'profile': True}`)
         more_replacements: `dict`
             Apply custom replacements before calculating gradients
 
         Returns
         -------
-        `aesara.function`
+        `pytensor.function`
         """
         if fn_kwargs is None:
             fn_kwargs = {}
@@ -362,13 +387,14 @@ class ObjectiveFunction:
             more_replacements=more_replacements,
             total_grad_norm_constraint=total_grad_norm_constraint,
         )
+        seed = self.approx.rng.randint(2**30, dtype=np.int64)
         if score:
-            step_fn = compile_pymc([], updates.loss, updates=updates, **fn_kwargs)
+            step_fn = compile_pymc([], updates.loss, updates=updates, random_seed=seed, **fn_kwargs)
         else:
-            step_fn = compile_pymc([], [], updates=updates, **fn_kwargs)
+            step_fn = compile_pymc([], [], updates=updates, random_seed=seed, **fn_kwargs)
         return step_fn
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def score_function(
         self, sc_n_mc=None, more_replacements=None, fn_kwargs=None
     ):  # pragma: no cover
@@ -381,11 +407,11 @@ class ObjectiveFunction:
         more_replacements:
             Apply custom replacements before compiling a function
         fn_kwargs: `dict`
-            arbitrary kwargs passed to `aesara.function`
+            arbitrary kwargs passed to `pytensor.function`
 
         Returns
         -------
-        aesara.function
+        pytensor.function
         """
         if fn_kwargs is None:
             fn_kwargs = {}
@@ -394,9 +420,10 @@ class ObjectiveFunction:
         if more_replacements is None:
             more_replacements = {}
         loss = self(sc_n_mc, more_replacements=more_replacements)
-        return compile_pymc([], loss, **fn_kwargs)
+        seed = self.approx.rng.randint(2**30, dtype=np.int64)
+        return compile_pymc([], loss, random_seed=seed, **fn_kwargs)
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def __call__(self, nmc, **kwargs):
         if "more_tf_params" in kwargs:
             m = -1.0
@@ -429,10 +456,6 @@ class Operator:
 
     def __init__(self, approx):
         self.approx = approx
-        if not self.supports_aevb and approx.has_local:
-            raise AEVBInferenceError(
-                "%s does not support AEVB, " "please change inference method" % self
-            )
         if self.require_logq and not approx.has_logq:
             raise ExplicitInferenceError(
                 "%s requires logq, but %s does not implement it"
@@ -508,7 +531,7 @@ def collect_shared_to_list(params):
         return list(
             t[1]
             for t in sorted(params.items(), key=lambda t: t[0])
-            if isinstance(t[1], aesara.compile.SharedVariable)
+            if isinstance(t[1], pytensor.compile.SharedVariable)
         )
     elif params is None:
         return []
@@ -560,12 +583,6 @@ class Group(WithMemoization):
         Random seed for underlying random generator
     model :
         PyMC Model
-    local: bool
-        Indicates whether this group is local. Cannot be passed without `params`.
-        Such group should have only one variable
-    rowwise: bool
-        Indicates whether this group is independently parametrized over first dim.
-        Such group should have only one variable
     options: dict
         Special options for the group
     kwargs: Other kwargs for the group
@@ -574,17 +591,6 @@ class Group(WithMemoization):
     -----
     Group instance/class has some important constants:
 
-    -   **supports_batched**
-        Determines whether such variational family can be used for AEVB or rowwise approx.
-
-        AEVB approx is such approx that somehow depends on input data. It can be treated
-        as conditional distribution. You can see more about in the corresponding paper
-        mentioned in references.
-
-        Rowwise mode is a special case approximation that treats every 'row', of a tensor as
-        independent from each other. Some distributions can't do that by
-        definition e.g. :class:`Empirical` that consists of particles only.
-
     -   **has_logq**
         Tells that distribution is defined explicitly
 
@@ -592,7 +598,6 @@ class Group(WithMemoization):
 
     Examples
     --------
-
     **Basic Initialization**
 
     :class:`Group` is a factory class. You do not need to call every ApproximationGroup explicitly.
@@ -624,101 +629,6 @@ class Group(WithMemoization):
     -   `{'mu', 'L_tril'}`: :class:`FullRankGroup`
 
     -   `{'histogram'}`: :class:`EmpiricalGroup`
-
-    -   `{0, 1, 2, 3, ..., k-1}`: :class:`NormalizingFlowGroup` of depth `k`
-
-        NormalizingFlows have other parameters than ordinary groups and should be
-        passed as nested dicts with the following keys:
-
-        -   `{'u', 'w', 'b'}`: :class:`PlanarFlow`
-
-        -   `{'a', 'b', 'z_ref'}`: :class:`RadialFlow`
-
-        -   `{'loc'}`: :class:`LocFlow`
-
-        -   `{'rho'}`: :class:`ScaleFlow`
-
-        -   `{'v'}`: :class:`HouseholderFlow`
-
-        Note that all integer keys should be present in the dictionary. An example
-        of NormalizingFlow initialization can be found below.
-
-    **Using AEVB**
-
-    Autoencoding variational Bayes is a powerful tool to get conditional :math:`q(\lambda|X)` distribution
-    on latent variables. It is well supported by PyMC and all you need is to provide a dictionary
-    with well shaped variational parameters, the correct approximation will be autoselected as mentioned
-    in section above. However we have some implementation restrictions in AEVB. They require autoencoded
-    variable to have first dimension as *batch* dimension and other dimensions should stay fixed.
-    With this assumptions it is possible to generalize all variational approximation families as
-    batched approximations that have flexible parameters and leading axis.
-
-    Only single variable local group is supported. Params are required.
-
-    >>> # for mean field
-    >>> group = Group([latent3], params=dict(mu=my_mu, rho=my_rho), local=True)
-    >>> # or for full rank
-    >>> group = Group([latent3], params=dict(mu=my_mu, L_tril=my_L_tril), local=True)
-
-    -   An Approximation class is selected automatically based on the keys in dict.
-
-    -   `my_mu` and `my_rho` are usually estimated with neural network or function approximator.
-
-    **Using Row-Wise Group**
-
-    Batch groups have independent row wise approximations, thus using batched
-    mean field will give no effect. It is more interesting if you want each row of a matrix
-    to be parametrized independently with normalizing flow or full rank gaussian.
-
-    To tell :class:`Group` that group is batched you need set `batched` kwarg as `True`.
-    Only single variable group is allowed due to implementation details.
-
-    >>> group = Group([latent3], vfam='fr', rowwise=True) # 'fr' is alias for 'full_rank'
-
-    The resulting approximation for this variable will have the following structure
-
-    .. math::
-
-        latent3_{i, \dots} \sim \mathcal{N}(\mu_i, \Sigma_i) \forall i
-
-    **Note**: Using rowwise and user-parametrized approximation is ok, but
-    shape should be checked beforehand, it is impossible to infer it by PyMC
-
-    **Normalizing Flow Group**
-
-    In case you use simple initialization pattern using `vfam` you'll not meet any changes.
-    Passing flow formula to `vfam` you'll get correct flow parametrization for group
-
-    .. code:: python
-
-        >>> group = Group([latent3], vfam='scale-hh*5-radial*4-loc')
-
-    **Note**: Consider passing location flow as the last one and scale as the first one for stable inference.
-
-    Rowwise normalizing flow is supported as well
-
-    .. code:: python
-
-        >>> group = Group([latent3], vfam='scale-hh*2-radial-loc', rowwise=True)
-
-    Custom parameters for normalizing flow can be a real trouble for the first time.
-    They have quite different format from the rest variational families.
-
-
-    .. code:: python
-
-        >>> # int is used as key, it also tells the flow position
-        ... flow_params = {
-        ...     # `rho` parametrizes scale flow, softplus is used to map (-inf; inf) -> (0, inf)
-        ...     0: dict(rho=my_scale),
-        ...     1: dict(v=my_v1),  # Householder Flow, `v` is parameter name from the original paper
-        ...     2: dict(v=my_v2),  # do not miss any number in dict, or else error is raised
-        ...     3: dict(a=my_a, b=my_b, z_ref=my_z_ref),  # Radial flow
-        ...     4: dict(loc=my_loc)  # Location Flow
-        ... }
-        ... group = Group([latent3], params=flow_params)
-        ... # local=True can be added in case you do AEVB inference
-        ... group = Group([latent3], params=flow_params, local=True)
 
     **Delayed Initialization**
 
@@ -760,7 +670,6 @@ class Group(WithMemoization):
     input = None
 
     # defined by approximation
-    supports_batched = True
     has_logq = True
 
     # some important defaults
@@ -768,11 +677,11 @@ class Group(WithMemoization):
     initial_dist_map = 0.0
 
     # for handy access using class methods
-    __param_spec__ = dict()
+    __param_spec__: dict = dict()
     short_name = ""
-    alias_names = frozenset()
-    __param_registry = dict()
-    __name_registry = dict()
+    alias_names: frozenset[str] = frozenset()
+    __param_registry: dict[frozenset, Any] = dict()
+    __name_registry: dict[str, Any] = dict()
 
     @classmethod
     def register(cls, sbcls):
@@ -789,8 +698,6 @@ class Group(WithMemoization):
 
     @classmethod
     def group_for_params(cls, params):
-        if pm.variational.flows.seems_like_flow_params(params):
-            return pm.variational.approximations.NormalizingFlowGroup
         if frozenset(params) not in cls.__param_registry:
             raise KeyError(
                 "No such group for the following params: {!r}, "
@@ -800,8 +707,6 @@ class Group(WithMemoization):
 
     @classmethod
     def group_for_short_name(cls, name):
-        if pm.variational.flows.seems_like_formula(name):
-            return pm.variational.approximations.NormalizingFlowGroup
         if name.lower() not in cls.__name_registry:
             raise KeyError(
                 "No such group: {!r}, "
@@ -829,24 +734,16 @@ class Group(WithMemoization):
         params=None,
         random_seed=None,
         model=None,
-        local=False,
-        rowwise=False,
         options=None,
         **kwargs,
     ):
-        if local and not self.supports_batched:
-            raise LocalGroupError("%s does not support local groups" % self.__class__)
-        if local and rowwise:
-            raise LocalGroupError("%s does not support local grouping in rowwise mode")
         if isinstance(vfam, str):
             vfam = vfam.lower()
         if options is None:
             options = dict()
         self.options = options
         self._vfam = vfam
-        self._local = local
-        self._batched = rowwise
-        self._rng = at_rng(random_seed)
+        self.rng = np.random.RandomState(random_seed)
         model = modelcontext(model)
         self.model = model
         self.group = group
@@ -867,13 +764,10 @@ class Group(WithMemoization):
             jitter_rvs={},
             return_transformed=True,
         )
-        start = ipfn(self.model.rng_seeder.randint(2**30, dtype=np.int64))
+        start = ipfn(self.rng.randint(2**30, dtype=np.int64))
         group_vars = {self.model.rvs_to_values[v].name for v in self.group}
         start = {k: v for k, v in start.items() if k in group_vars}
-        if self.batched:
-            start = start[self.group[0].name][0]
-        else:
-            start = DictToArrayBijection.map(start).data
+        start = DictToArrayBijection.map(start).data
         return start
 
     @classmethod
@@ -913,11 +807,7 @@ class Group(WithMemoization):
         spec = self.get_param_spec_for(d=self.ddim, **kwargs.pop("spec_kw", {}))
         for name, param in self.user_params.items():
             shape = spec[name]
-            if self.local:
-                shape = (-1,) + shape
-            elif self.batched:
-                shape = (self.bdim,) + shape
-            self._user_params[name] = at.as_tensor(param).reshape(shape)
+            self._user_params[name] = pt.as_tensor(param).reshape(shape)
         return True
 
     def _initial_type(self, name):
@@ -931,10 +821,7 @@ class Group(WithMemoization):
         -------
         tensor
         """
-        if self.batched:
-            return at.tensor3(name)
-        else:
-            return at.matrix(name)
+        return pt.matrix(name)
 
     def _input_type(self, name):
         R"""*Dev* - input type with given name. The correct type depends on `self.batched`
@@ -947,29 +834,15 @@ class Group(WithMemoization):
         -------
         tensor
         """
-        if self.batched:
-            return at.matrix(name)
-        else:
-            return at.vector(name)
+        return pt.vector(name)
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def __init_group__(self, group):
         if not group:
             raise GroupError("Got empty group")
-        if self.local:
-            raise NotImplementedInference("Local inferene aka AEVB is not supported in v4")
-        if self.batched:
-            raise NotImplementedInference("Batched inferene is not supported in v4")
         if self.group is None:
             # delayed init
             self.group = group
-        if self.batched and len(group) > 1:
-            if self.local:  # better error message
-                raise LocalGroupError("Local groups with more than 1 variable are not supported")
-            else:
-                raise BatchedGroupError(
-                    "Batched groups with more than 1 variable are not supported"
-                )
         self.symbolic_initial = self._initial_type(
             self.__class__.__name__ + "_symbolic_initial_tensor"
         )
@@ -978,7 +851,7 @@ class Group(WithMemoization):
         # so I have to to it by myself
 
         # 1) we need initial point (transformed space)
-        model_initial_point = self.model.compute_initial_point(0)
+        model_initial_point = self.model.initial_point(0)
         # 2) we'll work with a single group, a subset of the model
         # here we need to create a mapping to replace value_vars with slices from the approximation
         start_idx = 0
@@ -988,21 +861,8 @@ class Group(WithMemoization):
             # 3) This is the way to infer shape and dtype of the variable
             value_var = self.model.rvs_to_values[var]
             test_var = model_initial_point[value_var.name]
-            if self.batched:
-                # Leave a more complicated case for future work
-                if var.ndim < 1:
-                    if self.local:
-                        raise LocalGroupError("Local variable should not be scalar")
-                    else:
-                        raise BatchedGroupError("Batched variable should not be scalar")
-                size = test_var[0].size
-                if self.local:
-                    shape = (-1,) + test_var.shape[1:]
-                else:
-                    shape = test_var.shape
-            else:
-                shape = test_var.shape
-                size = test_var.size
+            shape = test_var.shape
+            size = test_var.size
             dtype = test_var.dtype
             vr = self.input[..., start_idx : start_idx + size].reshape(shape).astype(dtype)
             vr.name = value_var.name + "_vi_replacement"
@@ -1018,9 +878,6 @@ class Group(WithMemoization):
     def _finalize_init(self):
         """*Dev* - clean up after init"""
         del self._kwargs
-
-    local = property(lambda self: self._local)
-    batched = property(lambda self: self._local or self._batched)
 
     @property
     def params_dict(self):
@@ -1054,29 +911,11 @@ class Group(WithMemoization):
         -------
         shape vector
         """
-        if self.batched:
-            bdim = at.as_tensor(self.bdim)
-            bdim = aesara.clone_replace(bdim, more_replacements)
-            return at.stack([size, bdim, dim])
-        else:
-            return at.stack([size, dim])
-
-    @node_property
-    def bdim(self):
-        if not self.local:
-            if self.batched:
-                return next(iter(self.ordering.values()))[2][0]
-            else:
-                return 1
-        else:
-            return next(iter(self.params_dict.values())).shape[0]
+        return pt.stack([size, dim])
 
     @node_property
     def ndim(self):
-        if self.batched:
-            return self.ordering.size * self.bdim
-        else:
-            return self.ddim
+        return self.ddim
 
     @property
     def ddim(self):
@@ -1113,18 +952,18 @@ class Group(WithMemoization):
             deterministic = np.int8(deterministic)
         dim, dist_name, dist_map = (self.ddim, self.initial_dist_name, self.initial_dist_map)
         dtype = self.symbolic_initial.dtype
-        dim = at.as_tensor(dim)
-        size = at.as_tensor(size)
+        dim = pt.as_tensor(dim)
+        size = pt.as_tensor(size)
         shape = self._new_initial_shape(size, dim, more_replacements)
         # apply optimizations if possible
         if not isinstance(deterministic, Variable):
             if deterministic:
-                return at.ones(shape, dtype) * dist_map
+                return pt.ones(shape, dtype) * dist_map
             else:
-                return getattr(self._rng, dist_name)(size=shape)
+                return getattr(pt.random, dist_name)(size=shape)
         else:
-            sample = getattr(self._rng, dist_name)(size=shape)
-            initial = at.switch(deterministic, at.ones(shape, dtype) * dist_map, sample)
+            sample = getattr(pt.random, dist_name)(size=shape)
+            initial = pt.switch(deterministic, pt.ones(shape, dtype) * dist_map, sample)
             return initial
 
     @node_property
@@ -1141,23 +980,17 @@ class Group(WithMemoization):
         """
         raise NotImplementedError
 
-    @node_property
-    def symbolic_random2d(self):
-        """*Dev* - `self.symbolic_random` flattened to matrix"""
-        if self.batched:
-            return self.symbolic_random.flatten(2)
-        else:
-            return self.symbolic_random
-
-    @aesara.config.change_flags(compute_test_value="off")
-    def set_size_and_deterministic(self, node, s, d, more_replacements=None):
+    @pytensor.config.change_flags(compute_test_value="off")
+    def set_size_and_deterministic(
+        self, node: Variable, s, d: bool, more_replacements: dict = None
+    ) -> list[Variable]:
         """*Dev* - after node is sampled via :func:`symbolic_sample_over_posterior` or
         :func:`symbolic_single_sample` new random generator can be allocated and applied to node
 
         Parameters
         ----------
         node: :class:`Variable`
-            Aesara node with symbolically applied VI replacements
+            PyTensor node with symbolically applied VI replacements
         s: scalar
             desired number of samples
         d: bool or int
@@ -1169,14 +1002,19 @@ class Group(WithMemoization):
         -------
         :class:`Variable` with applied replacements, ready to use
         """
+
         flat2rand = self.make_size_and_deterministic_replacements(s, d, more_replacements)
-        node_out = aesara.clone_replace(node, flat2rand)
+        node_out = graph_replace(node, flat2rand, strict=False)
+        assert not (
+            set(makeiter(self.input)) & set(pytensor.graph.graph_inputs(makeiter(node_out)))
+        )
         try_to_set_test_value(node, node_out, s)
+        assert self.symbolic_random not in set(pytensor.graph.graph_inputs(makeiter(node_out)))
         return node_out
 
     def to_flat_input(self, node):
         """*Dev* - replace vars with flattened view stored in `self.inputs`"""
-        return aesara.clone_replace(node, self.replacements)
+        return graph_replace(node, self.replacements, strict=False)
 
     def symbolic_sample_over_posterior(self, node):
         """*Dev* - performs sampling of node applying independent samples from posterior each time.
@@ -1184,12 +1022,15 @@ class Group(WithMemoization):
         """
         node = self.to_flat_input(node)
         random = self.symbolic_random.astype(self.symbolic_initial.dtype)
-        random = at.patternbroadcast(random, self.symbolic_initial.broadcastable)
+        random = pt.specify_shape(random, self.symbolic_initial.type.shape)
 
-        def sample(post, node):
-            return aesara.clone_replace(node, {self.input: post})
+        def sample(post, *_):
+            return graph_replace(node, {self.input: post}, strict=False)
 
-        nodes, _ = aesara.scan(sample, random, non_sequences=[node])
+        nodes, _ = pytensor.scan(
+            sample, random, non_sequences=_known_scan_ignored_inputs(makeiter(random))
+        )
+        assert self.input not in set(pytensor.graph.graph_inputs(makeiter(nodes)))
         return nodes
 
     def symbolic_single_sample(self, node):
@@ -1199,7 +1040,7 @@ class Group(WithMemoization):
         """
         node = self.to_flat_input(node)
         random = self.symbolic_random.astype(self.symbolic_initial.dtype)
-        return aesara.clone_replace(node, {self.input: random[0]})
+        return graph_replace(node, {self.input: random[0]}, strict=False)
 
     def make_size_and_deterministic_replacements(self, s, d, more_replacements=None):
         """*Dev* - creates correct replacements for initial depending on
@@ -1219,15 +1060,31 @@ class Group(WithMemoization):
         dict with replacements for initial
         """
         initial = self._new_initial(s, d, more_replacements)
-        initial = at.patternbroadcast(initial, self.symbolic_initial.broadcastable)
+        initial = pt.specify_shape(initial, self.symbolic_initial.type.shape)
+        # The static shape of initial may be more precise than self.symbolic_initial,
+        # and reveal previously unknown broadcastable dimensions. We have to mask those again.
+        if initial.type.broadcastable != self.symbolic_initial.type.broadcastable:
+            unbroadcast_axes = (
+                i for i, b in enumerate(self.symbolic_initial.type.broadcastable) if not b
+            )
+            initial = unbroadcast(initial, *unbroadcast_axes)
         if more_replacements:
-            initial = aesara.clone_replace(initial, more_replacements)
+            initial = graph_replace(initial, more_replacements, strict=False)
         return {self.symbolic_initial: initial}
 
     @node_property
     def symbolic_normalizing_constant(self):
         """*Dev* - normalizing constant for `self.logq`, scales it to `minibatch_size` instead of `total_size`"""
-        t = self.to_flat_input(at.max([v.tag.scaling for v in self.group]))
+        t = self.to_flat_input(
+            pt.max(
+                [
+                    get_scaling(v.owner.inputs[1:], v.shape)
+                    for v in self.group
+                    if isinstance(v.owner.op, MinibatchRandomVariable)
+                ]
+                + [1.0]  # To avoid empty max
+            )
+        )
         t = self.symbolic_single_sample(t)
         return pm.floatX(t)
 
@@ -1242,13 +1099,7 @@ class Group(WithMemoization):
     @node_property
     def symbolic_logq(self):
         """*Dev* - correctly scaled `self.symbolic_logq_not_scaled`"""
-        if self.local:
-            s = self.group[0].tag.scaling
-            s = self.to_flat_input(s)
-            s = self.symbolic_single_sample(s)
-            return self.symbolic_logq_not_scaled * s
-        else:
-            return self.symbolic_logq_not_scaled
+        return self.symbolic_logq_not_scaled
 
     @node_property
     def logq(self):
@@ -1265,23 +1116,50 @@ class Group(WithMemoization):
             shp = "undefined"
         else:
             shp = str(self.ddim)
-            if self.local:
-                shp = "None, " + shp
-            elif self.batched:
-                shp = str(self.bdim) + ", " + shp
         return f"{self.__class__.__name__}[{shp}]"
 
     @node_property
-    def std(self):
-        raise NotImplementedError
+    def std(self) -> pt.TensorVariable:
+        """Standard deviation of the latent variables as an unstructured 1-dimensional tensor variable"""
+        raise NotImplementedError()
 
     @node_property
-    def cov(self):
-        raise NotImplementedError
+    def cov(self) -> pt.TensorVariable:
+        """Covariance between the latent variables as an unstructured 2-dimensional tensor variable"""
+        raise NotImplementedError()
 
     @node_property
-    def mean(self):
-        raise NotImplementedError
+    def mean(self) -> pt.TensorVariable:
+        """Mean of the latent variables as an unstructured 1-dimensional tensor variable"""
+        raise NotImplementedError()
+
+    def var_to_data(self, shared: pt.TensorVariable) -> xarray.Dataset:
+        """Takes a flat 1-dimensional tensor variable and maps it to an xarray data set based on the information in
+        `self.ordering`.
+        """
+        # This is somewhat similar to `DictToArrayBijection.rmap`, which doesn't work here since we don't have
+        # `RaveledVars` and need to take the information from `self.ordering` instead
+        shared_nda = shared.eval()
+        result = dict()
+        for name, s, shape, dtype in self.ordering.values():
+            dims = self.model.named_vars_to_dims.get(name, None)
+            if dims is not None:
+                coords = {d: np.array(self.model.coords[d]) for d in dims}
+            else:
+                coords = None
+            values = shared_nda[s].reshape(shape).astype(dtype)
+            result[name] = xarray.DataArray(values, coords=coords, dims=dims, name=name)
+        return xarray.Dataset(result)
+
+    @property
+    def mean_data(self) -> xarray.Dataset:
+        """Mean of the latent variables as an xarray Dataset"""
+        return self.var_to_data(self.mean)
+
+    @property
+    def std_data(self) -> xarray.Dataset:
+        """Standard deviation of the latent variables as an xarray Dataset"""
+        return self.var_to_data(self.std)
 
 
 group_for_params = Group.group_for_params
@@ -1307,11 +1185,7 @@ class Approximation(WithMemoization):
 
         -   :class:`MeanField`
         -   :class:`FullRank`
-        -   :class:`NormalizingFlow`
         -   :class:`Empirical`
-
-    Single group accepts `local_rv` keyword with dict mapping PyMC variables
-    to their local Group parameters dict
 
     See Also
     --------
@@ -1319,7 +1193,7 @@ class Approximation(WithMemoization):
     """
 
     def __init__(self, groups, model=None):
-        self._scale_cost_to_minibatch = aesara.shared(np.int8(1))
+        self._scale_cost_to_minibatch = pytensor.shared(np.int8(1))
         model = modelcontext(model)
         if not model.free_RVs:
             raise TypeError("Model does not have an free RVs")
@@ -1337,11 +1211,13 @@ class Approximation(WithMemoization):
                     raise GroupError("Found duplicates in groups")
                 seen.update(g.group)
                 self.groups.append(g)
-        if set(model.free_RVs) - seen:
+        # List iteration to preserve order for reproducibility between runs
+        unseen_free_RVs = [var for var in model.free_RVs if var not in seen]
+        if unseen_free_RVs:
             if rest is None:
                 raise GroupError("No approximation is specified for the rest variables")
             else:
-                rest.__init_group__(list(set(model.free_RVs) - seen))
+                rest.__init_group__(unseen_free_RVs)
                 self.groups.append(rest)
         self.model = model
 
@@ -1349,17 +1225,8 @@ class Approximation(WithMemoization):
     def has_logq(self):
         return all(self.collect("has_logq"))
 
-    def collect(self, item, part="total"):
-        if part == "total":
-            return [getattr(g, item) for g in self.groups]
-        elif part == "local":
-            return [getattr(g, item) for g in self.groups if g.local]
-        elif part == "global":
-            return [getattr(g, item) for g in self.groups if not g.local]
-        elif part == "batched":
-            return [getattr(g, item) for g in self.groups if g.batched]
-        else:
-            raise ValueError("unknown part %s, expected {'local', 'global', 'total', 'batched'}")
+    def collect(self, item):
+        return [getattr(g, item) for g in self.groups]
 
     inputs = property(lambda self: self.collect("input"))
     symbolic_randoms = property(lambda self: self.collect("symbolic_random"))
@@ -1378,22 +1245,26 @@ class Approximation(WithMemoization):
         """*Dev* - normalizing constant for `self.logq`, scales it to `minibatch_size` instead of `total_size`.
         Here the effect is controlled by `self.scale_cost_to_minibatch`
         """
-        t = at.max(
+        t = pt.max(
             self.collect("symbolic_normalizing_constant")
-            + [var.tag.scaling for var in self.model.observed_RVs]
+            + [
+                get_scaling(obs.owner.inputs[1:], obs.shape)
+                for obs in self.model.observed_RVs
+                if isinstance(obs.owner.op, MinibatchRandomVariable)
+            ]
         )
-        t = at.switch(self._scale_cost_to_minibatch, t, at.constant(1, dtype=t.dtype))
+        t = pt.switch(self._scale_cost_to_minibatch, t, pt.constant(1, dtype=t.dtype))
         return pm.floatX(t)
 
     @node_property
     def symbolic_logq(self):
         """*Dev* - collects `symbolic_logq` for all groups"""
-        return at.add(*self.collect("symbolic_logq"))
+        return pt.add(*self.collect("symbolic_logq"))
 
     @node_property
     def logq(self):
         """*Dev* - collects `logQ` for all groups"""
-        return at.add(*self.collect("logq"))
+        return pt.add(*self.collect("logq"))
 
     @node_property
     def logq_norm(self):
@@ -1402,63 +1273,63 @@ class Approximation(WithMemoization):
 
     @node_property
     def _sized_symbolic_varlogp_and_datalogp(self):
-        """*Dev* - computes sampled prior term from model via `aesara.scan`"""
+        """*Dev* - computes sampled prior term from model via `pytensor.scan`"""
         varlogp_s, datalogp_s = self.symbolic_sample_over_posterior(
-            [self.model.varlogpt, self.model.datalogpt]
+            [self.model.varlogp, self.model.datalogp]
         )
         return varlogp_s, datalogp_s  # both shape (s,)
 
     @node_property
     def sized_symbolic_varlogp(self):
-        """*Dev* - computes sampled prior term from model via `aesara.scan`"""
+        """*Dev* - computes sampled prior term from model via `pytensor.scan`"""
         return self._sized_symbolic_varlogp_and_datalogp[0]  # shape (s,)
 
     @node_property
     def sized_symbolic_datalogp(self):
-        """*Dev* - computes sampled data term from model via `aesara.scan`"""
+        """*Dev* - computes sampled data term from model via `pytensor.scan`"""
         return self._sized_symbolic_varlogp_and_datalogp[1]  # shape (s,)
 
     @node_property
     def sized_symbolic_logp(self):
-        """*Dev* - computes sampled logP from model via `aesara.scan`"""
+        """*Dev* - computes sampled logP from model via `pytensor.scan`"""
         return self.sized_symbolic_varlogp + self.sized_symbolic_datalogp  # shape (s,)
 
     @node_property
     def logp(self):
-        """*Dev* - computes :math:`E_{q}(logP)` from model via `aesara.scan` that can be optimized later"""
+        """*Dev* - computes :math:`E_{q}(logP)` from model via `pytensor.scan` that can be optimized later"""
         return self.varlogp + self.datalogp
 
     @node_property
     def varlogp(self):
-        """*Dev* - computes :math:`E_{q}(prior term)` from model via `aesara.scan` that can be optimized later"""
+        """*Dev* - computes :math:`E_{q}(prior term)` from model via `pytensor.scan` that can be optimized later"""
         return self.sized_symbolic_varlogp.mean(0)
 
     @node_property
     def datalogp(self):
-        """*Dev* - computes :math:`E_{q}(data term)` from model via `aesara.scan` that can be optimized later"""
+        """*Dev* - computes :math:`E_{q}(data term)` from model via `pytensor.scan` that can be optimized later"""
         return self.sized_symbolic_datalogp.mean(0)
 
     @node_property
     def _single_symbolic_varlogp_and_datalogp(self):
-        """*Dev* - computes sampled prior term from model via `aesara.scan`"""
-        varlogp, datalogp = self.symbolic_single_sample([self.model.varlogpt, self.model.datalogpt])
+        """*Dev* - computes sampled prior term from model via `pytensor.scan`"""
+        varlogp, datalogp = self.symbolic_single_sample([self.model.varlogp, self.model.datalogp])
         return varlogp, datalogp
 
     @node_property
     def single_symbolic_varlogp(self):
-        """*Dev* - for single MC sample estimate of :math:`E_{q}(prior term)` `aesara.scan`
+        """*Dev* - for single MC sample estimate of :math:`E_{q}(prior term)` `pytensor.scan`
         is not needed and code can be optimized"""
         return self._single_symbolic_varlogp_and_datalogp[0]
 
     @node_property
     def single_symbolic_datalogp(self):
-        """*Dev* - for single MC sample estimate of :math:`E_{q}(data term)` `aesara.scan`
+        """*Dev* - for single MC sample estimate of :math:`E_{q}(data term)` `pytensor.scan`
         is not needed and code can be optimized"""
         return self._single_symbolic_varlogp_and_datalogp[1]
 
     @node_property
     def single_symbolic_logp(self):
-        """*Dev* - for single MC sample estimate of :math:`E_{q}(logP)` `aesara.scan`
+        """*Dev* - for single MC sample estimate of :math:`E_{q}(logP)` `pytensor.scan`
         is not needed and code can be optimized"""
         return self.single_symbolic_datalogp + self.single_symbolic_varlogp
 
@@ -1509,7 +1380,7 @@ class Approximation(WithMemoization):
         flat2rand.update(more_replacements)
         return flat2rand
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def set_size_and_deterministic(self, node, s, d, more_replacements=None):
         """*Dev* - after node is sampled via :func:`symbolic_sample_over_posterior` or
         :func:`symbolic_single_sample` new random generator can be allocated and applied to node
@@ -1517,7 +1388,7 @@ class Approximation(WithMemoization):
         Parameters
         ----------
         node: :class:`Variable`
-            Aesara node with symbolically applied VI replacements
+            PyTensor node with symbolically applied VI replacements
         s: scalar
             desired number of samples
         d: bool or int
@@ -1532,27 +1403,31 @@ class Approximation(WithMemoization):
         _node = node
         optimizations = self.get_optimization_replacements(s, d)
         flat2rand = self.make_size_and_deterministic_replacements(s, d, more_replacements)
-        node = aesara.clone_replace(node, optimizations)
-        node = aesara.clone_replace(node, flat2rand)
+        node = graph_replace(node, optimizations, strict=False)
+        node = graph_replace(node, flat2rand, strict=False)
+        assert not (set(self.symbolic_randoms) & set(pytensor.graph.graph_inputs(makeiter(node))))
         try_to_set_test_value(_node, node, s)
         return node
 
     def to_flat_input(self, node, more_replacements=None):
         """*Dev* - replace vars with flattened view stored in `self.inputs`"""
         more_replacements = more_replacements or {}
-        node = aesara.clone_replace(node, more_replacements)
-        return aesara.clone_replace(node, self.replacements)
+        node = graph_replace(node, more_replacements, strict=False)
+        return graph_replace(node, self.replacements, strict=False)
 
     def symbolic_sample_over_posterior(self, node, more_replacements=None):
         """*Dev* - performs sampling of node applying independent samples from posterior each time.
         Note that it is done symbolically and this node needs :func:`set_size_and_deterministic` call
         """
-        node = self.to_flat_input(node, more_replacements=more_replacements)
+        node = self.to_flat_input(node)
 
         def sample(*post):
-            return aesara.clone_replace(node, dict(zip(self.inputs, post)))
+            return graph_replace(node, dict(zip(self.inputs, post)), strict=False)
 
-        nodes, _ = aesara.scan(sample, self.symbolic_randoms)
+        nodes, _ = pytensor.scan(
+            sample, self.symbolic_randoms, non_sequences=_known_scan_ignored_inputs(makeiter(node))
+        )
+        assert not (set(self.inputs) & set(pytensor.graph.graph_inputs(makeiter(nodes))))
         return nodes
 
     def symbolic_single_sample(self, node, more_replacements=None):
@@ -1563,11 +1438,11 @@ class Approximation(WithMemoization):
         node = self.to_flat_input(node, more_replacements=more_replacements)
         post = [v[0] for v in self.symbolic_randoms]
         inp = self.inputs
-        return aesara.clone_replace(node, dict(zip(inp, post)))
+        return graph_replace(node, dict(zip(inp, post)), strict=False)
 
     def get_optimization_replacements(self, s, d):
         """*Dev* - optimizations for logP. If sample size is static and equal to 1:
-        then `aesara.scan` MC estimate is replaced with single sample without call to `aesara.scan`.
+        then `pytensor.scan` MC estimate is replaced with single sample without call to `pytensor.scan`.
         """
         repl = collections.OrderedDict()
         # avoid scan if size is constant and equal to one
@@ -1576,13 +1451,13 @@ class Approximation(WithMemoization):
             repl[self.datalogp] = self.single_symbolic_datalogp
         return repl
 
-    @aesara.config.change_flags(compute_test_value="off")
+    @pytensor.config.change_flags(compute_test_value="off")
     def sample_node(self, node, size=None, deterministic=False, more_replacements=None):
         """Samples given node or nodes over shared posterior
 
         Parameters
         ----------
-        node: Aesara Variables (or Aesara expressions)
+        node: PyTensor Variables (or PyTensor expressions)
         size: None or scalar
             number of samples
         more_replacements: `dict`
@@ -1597,10 +1472,10 @@ class Approximation(WithMemoization):
         """
         node_in = node
         if more_replacements:
-            node = aesara.clone_replace(node, more_replacements)
+            node = graph_replace(node, more_replacements, strict=False)
         if not isinstance(node, (list, tuple)):
             node = [node]
-        node, _ = rvs_to_value_vars(node, apply_transforms=True)
+        node = self.model.replace_rvs_by_values(node)
         if not isinstance(node_in, (list, tuple)):
             node = node[0]
         if size is None:
@@ -1612,7 +1487,7 @@ class Approximation(WithMemoization):
         return node_out
 
     def rslice(self, name):
-        """*Dev* - vectorized sampling for named random variable without call to `aesara.scan`.
+        """*Dev* - vectorized sampling for named random variable without call to `pytensor.scan`.
         This node still needs :func:`set_size_and_deterministic` to be evaluated
         """
 
@@ -1633,27 +1508,35 @@ class Approximation(WithMemoization):
 
     @node_property
     def sample_dict_fn(self):
-        s = at.iscalar()
+        s = pt.iscalar()
         names = [self.model.rvs_to_values[v].name for v in self.model.free_RVs]
         sampled = [self.rslice(name) for name in names]
         sampled = self.set_size_and_deterministic(sampled, s, 0)
         sample_fn = compile_pymc([s], sampled)
+        rng_nodes = find_rng_nodes(sampled)
 
-        def inner(draws=100):
+        def inner(draws=100, *, random_seed: SeedSequenceSeed = None):
+            if random_seed is not None:
+                reseed_rngs(rng_nodes, random_seed)
             _samples = sample_fn(draws)
+
             return {v_: s_ for v_, s_ in zip(names, _samples)}
 
         return inner
 
-    def sample(self, draws=500, return_inferencedata=True, **kwargs):
+    def sample(
+        self, draws=500, *, random_seed: RandomState = None, return_inferencedata=True, **kwargs
+    ):
         """Draw samples from variational posterior.
 
         Parameters
         ----------
-        draws: `int`
+        draws : int
             Number of random samples.
-        return_inferencedata: `bool`
-            Return trace in Arviz format
+        random_seed : int, RandomState or Generator, optional
+            Seed for the random number generator.
+        return_inferencedata : bool
+            Return trace in Arviz format.
 
         Returns
         -------
@@ -1663,7 +1546,9 @@ class Approximation(WithMemoization):
         # TODO: add tests for include_transformed case
         kwargs["log_likelihood"] = False
 
-        samples = self.sample_dict_fn(draws)  # type: dict
+        if random_seed is not None:
+            (random_seed,) = _get_seeds_per_chain(random_seed, 1)
+        samples: dict = self.sample_dict_fn(draws, random_seed=random_seed)
         points = ({name: records[i] for name, records in samples.items()} for i in range(draws))
 
         trace = NDArray(
@@ -1677,11 +1562,11 @@ class Approximation(WithMemoization):
         finally:
             trace.close()
 
-        trace = pm.sampling.MultiTrace([trace])
+        multi_trace = MultiTrace([trace])
         if not return_inferencedata:
-            return trace
+            return multi_trace
         else:
-            return pm.to_inference_data(trace, model=self.model, **kwargs)
+            return pm.to_inference_data(multi_trace, model=self.model, **kwargs)
 
     @property
     def ndim(self):
@@ -1691,21 +1576,9 @@ class Approximation(WithMemoization):
     def ddim(self):
         return sum(self.collect("ddim"))
 
-    @property
-    def has_local(self):
-        return any(self.collect("local"))
-
-    @property
-    def has_global(self):
-        return any(not c for c in self.collect("local"))
-
-    @property
-    def has_batched(self):
-        return any(not c for c in self.collect("batched"))
-
     @node_property
     def symbolic_random(self):
-        return at.concatenate(self.collect("symbolic_random2d"), axis=-1)
+        return pt.concatenate(self.collect("symbolic_random"), axis=-1)
 
     def __str__(self):
         if len(self.groups) < 5:
@@ -1726,7 +1599,7 @@ class Approximation(WithMemoization):
     def joint_histogram(self):
         if not self.all_histograms:
             raise VariationalInferenceError("%s does not consist of all Empirical approximations")
-        return at.concatenate(self.collect("histogram"), axis=-1)
+        return pt.concatenate(self.collect("histogram"), axis=-1)
 
     @property
     def params(self):
