@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -12,9 +12,9 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from aesara import tensor as at
+from pytensor import tensor as pt
 
-from pymc.aesaraf import floatX
+from pymc.pytensorf import floatX
 from pymc.variational.opvi import TestFunction
 
 __all__ = ["rbf"]
@@ -34,30 +34,30 @@ class Kernel(TestFunction):
 class RBF(Kernel):
     def __call__(self, X):
         XY = X.dot(X.T)
-        x2 = at.sum(X**2, axis=1).dimshuffle(0, "x")
-        X2e = at.repeat(x2, X.shape[0], axis=1)
+        x2 = pt.sum(X**2, axis=1).dimshuffle(0, "x")
+        X2e = pt.repeat(x2, X.shape[0], axis=1)
         H = X2e + X2e.T - 2.0 * XY
 
-        V = at.sort(H.flatten())
+        V = pt.sort(H.flatten())
         length = V.shape[0]
         # median distance
-        m = at.switch(
-            at.eq((length % 2), 0),
+        m = pt.switch(
+            pt.eq((length % 2), 0),
             # if even vector
-            at.mean(V[((length // 2) - 1) : ((length // 2) + 1)]),
+            pt.mean(V[((length // 2) - 1) : ((length // 2) + 1)]),
             # if odd vector
             V[length // 2],
         )
 
-        h = 0.5 * m / at.log(floatX(H.shape[0]) + floatX(1))
+        h = 0.5 * m / pt.log(floatX(H.shape[0]) + floatX(1))
 
         #  RBF
-        Kxy = at.exp(-H / h / 2.0)
+        Kxy = pt.exp(-H / h / 2.0)
 
         # Derivative
-        dxkxy = -at.dot(Kxy, X)
-        sumkxy = at.sum(Kxy, axis=-1, keepdims=True)
-        dxkxy = at.add(dxkxy, at.mul(X, sumkxy)) / h
+        dxkxy = -pt.dot(Kxy, X)
+        sumkxy = pt.sum(Kxy, axis=-1, keepdims=True)
+        dxkxy = pt.add(dxkxy, pt.mul(X, sumkxy)) / h
 
         return Kxy, dxkxy
 
