@@ -23,21 +23,21 @@ def find_testfiles():
     dp_repo = Path(__file__).parent.parent
     all_tests = {
         str(fp.relative_to(dp_repo)).replace(os.sep, "/")
-        for fp in (dp_repo / "pymc" / "tests").glob("**/test_*.py")
+        for fp in (dp_repo / "tests").glob("**/test_*.py")
     }
     _log.info("Found %i tests in total.", len(all_tests))
     return all_tests
 
 
 def from_yaml():
-    """Determins how often each test file is run per platform and floatX setting.
+    """Determines how often each test file is run per platform and floatX setting.
 
     An exception is raised if tests run multiple times with the same configuration.
     """
     # First collect the matrix definitions from testing workflows
     matrices = {}
-    for wf in ["pytest.yml", "arviz_compat.yml", "jaxtests.yml"]:
-        wfname = wf.strip(".yml")
+    for wf in ["tests.yml"]:
+        wfname = wf.rstrip(".yml")
         wfdef = yaml.safe_load(open(Path(".github", "workflows", wf)))
         for jobname, jobdef in wfdef["jobs"].items():
             matrix = jobdef.get("strategy", {}).get("matrix", {})
@@ -69,9 +69,28 @@ def from_yaml():
         for os_, floatX, subset in itertools.product(
             matrix["os"], matrix["floatx"], matrix["test-subset"]
         ):
-            testfiles = subset.split("\n")
-            ignored = {item.strip("--ignore=") for item in testfiles if item.startswith("--ignore")}
+            lines = [l for l in subset.split("\n") if l]
+            if "windows" in os_:
+                # Windows jobs need \ in line breaks within the test-subset!
+                # The following checks that these trailing \ are present in
+                # all items except the last.
+                if lines and lines[-1].endswith(" \\"):
+                    raise Exception(
+                        f"Last entry '{line}' in Windows test subset should end WITHOUT ' \\'."
+                    )
+                for line in lines[:-1]:
+                    if not line.endswith(" \\"):
+                        raise Exception(f"Missing ' \\' after '{line}' in Windows test-subset.")
+                lines = [line.rstrip(" \\") for line in lines]
+
+            # Unpack lines with >1 item
+            testfiles = []
+            for line in lines:
+                testfiles += line.split(" ")
+
+            ignored = {item[8:].lstrip(" =") for item in testfiles if item.startswith("--ignore")}
             included = {item for item in testfiles if item and not item.startswith("--ignore")}
+
             if ignored and not included:
                 # if no testfile is specified explicitly pytest runs all except the ignored ones
                 included = all_tests - ignored
@@ -86,9 +105,11 @@ def from_yaml():
     _log.info("Number of test runs (❌=0, ✅=once)\n%s", df.replace(0, "❌").replace(1, "✅"))
 
     if ignored_by_all:
-        _log.warning("%i tests are completely ignored:\n%s", len(ignored_by_all), ignored_by_all)
+        raise AssertionError(
+            f"{len(ignored_by_all)} tests are completely ignored:\n{ignored_by_all}"
+        )
     if run_multiple_times:
-        raise Exception(
+        raise AssertionError(
             f"{len(run_multiple_times)} tests are run multiple times with the same OS and floatX setting:\n{run_multiple_times}"
         )
     return
