@@ -31,7 +31,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 """
-Check requirements-dev.txt has been generated from conda-envs/environment-dev-py3*.yml
+Check requirements-dev.txt has been generated from conda-envs/environment-dev.yml
 
 This is intended to be used as a pre-commit hook, see `.pre-commit-config.yaml`.
 You can run it manually with `pre-commit run pip-from-conda --all`.
@@ -51,6 +51,7 @@ EXCLUDE = {
     "mkl-service",
     "numba",
     "python-graphviz",
+    "networkx",
     "blas",
     "jax",
 }
@@ -126,12 +127,12 @@ def main(conda_fname, pip_fname):
             raise ValueError(f"Unexpected dependency {dep}")
 
     header = (
-        f"# This file is auto-generated by scripts/generate_pip_deps_from_conda.py, "
+        "# This file is auto-generated by scripts/generate_pip_deps_from_conda.py, "
         "do not modify.\n# See that file for comments about the need/usage of each dependency.\n\n"
     )
     pip_content = header + "\n".join(sorted(pip_deps)) + "\n"
 
-    with open(pip_fname, "w") as pip_fd:
+    with open(pip_fname, "w", newline="") as pip_fd:
         pip_fd.write(pip_content)
 
 
