@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2023 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 A setup.py script to use setuptools, which gives egg goodness, etc.
 """
 
-from setuptools import setup
 
 with open("setup.py") as s:
     exec(s.read())
